### 2024-03-26 #6538 @kyleniemeyer
Hi @djs0109, the JOSS editorial board is going to take a closer look at your submission to see if this meets our scope as [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software). We'll get back to you within a week or two.
