### 2024-03-26 e2a6010
Proper deprecation exception for Quantumleap subscription.
