### 2024-03-26 2cd1c8a
fix: relationship will be removed when override entity
