### 2024-03-26 9d3280e
feat: add keyvalues support for post entity
