+++ filip/clients/ngsi_v2/cb.py
@@ -197,10 +197,11 @@ class ContextBrokerClient(BaseHttpClient):
     # Entity Operations
     def post_entity(
         self,
-        entity: ContextEntity,
+        entity: Union[ContextEntity, ContextEntityKeyValues],
         update: bool = False,
         patch: bool = False,
         override_attr_metadata: bool = True,
+        key_values: bool = False,
     ):
         """
         Function registers an Object with the NGSI Context Broker,
@@ -215,7 +216,7 @@ class ContextBrokerClient(BaseHttpClient):
         patch argument.
 
         Args:
-            entity (ContextEntity):
+            entity (ContextEntity/ContextEntityKeyValues):
                 Context Entity Object
             update (bool):
                 If the response.status_code is 422, whether the override and
@@ -227,13 +228,27 @@ class ContextBrokerClient(BaseHttpClient):
                 Only applies for patch equal to `True`.
                 Whether to override or append the attribute's metadata.
                 `True` for overwrite or `False` for update/append
-
+            key_values(bool):
+                By default False. If set to True, "options=keyValues" will
+                be included in params of  post request. The payload uses
+                the keyValues simplified entity representation, i.e.
+                ContextEntityKeyValues.
         """
         url = urljoin(self.base_url, "v2/entities")
         headers = self.headers.copy()
+        params = {}
+        options = []
+        if key_values:
+            assert isinstance(entity, ContextEntityKeyValues)
+            options.append("keyValues")
+        else:
+            assert isinstance(entity, ContextEntity)
+        if options:
+            params.update({'options': ",".join(options)})
         try:
             res = self.post(
-                url=url, headers=headers, json=entity.model_dump(exclude_none=True)
+                url=url, headers=headers, json=entity.model_dump(exclude_none=True),
+                params=params,
             )
             if res.ok:
                 self.logger.info("Entity successfully posted!")
@@ -242,11 +257,14 @@ class ContextBrokerClient(BaseHttpClient):
         except requests.RequestException as err:
             if update and err.response.status_code == 422:
                 return self.override_entity(
-                    entity=entity)
+                    entity=entity, key_values=key_values)
             if patch and err.response.status_code == 422:
-                return self.patch_entity(
-                    entity=entity, override_attr_metadata=override_attr_metadata
-                )
+                if not key_values:
+                    return self.patch_entity(
+                        entity=entity, override_attr_metadata=override_attr_metadata
+                    )
+                else:
+                    return self.update_entity_key_values(entity=entity)
             msg = f"Could not post entity {entity.id}"
             self.log_error(err=err, msg=msg)
             raise
@@ -812,7 +830,7 @@ class ContextBrokerClient(BaseHttpClient):
             "type": entity_type
         })
         entity = ContextEntityKeyValues(**entity_dict)
-        self.update_entity_key_value(entity=entity)
+        self.update_entity_key_values(entity=entity)
 
     def update_existing_entity_attributes(
             self,
@@ -879,7 +897,10 @@ class ContextBrokerClient(BaseHttpClient):
             self.log_error(err=err, msg=msg)
             raise
 
-    def override_entity(self, entity: ContextEntity):
+    def override_entity(self,
+                        entity: Union[ContextEntity, ContextEntityKeyValues],
+                        **kwargs
+                        ):
         """
         The request payload is an object representing the attributes to
         override the existing entity.
@@ -888,21 +909,25 @@ class ContextBrokerClient(BaseHttpClient):
             If you want to manipulate you should rather use patch_entity.
 
         Args:
-            entity (ContextEntity):
+            entity (ContextEntity or ContextEntityKeyValues):
         Returns:
             None
         """
         return self.replace_entity_attributes(entity_id=entity.id,
                                               entity_type=entity.type,
[…]
