### 2024-03-26 2298ca4
Testcase for endpoint, patch attribute of entity.
