### 2024-03-26 74cf079
chore: change default service_path to "/"
