+++ filip/models/base.py
@@ -103,7 +103,7 @@ class FiwareHeader(BaseModel):
     )
     service_path: str = Field(
         alias="fiware-servicepath",
-        default="",
+        default="/",
         description="Fiware service path",
         max_length=51,
     )
