### 2024-03-26 9700087
chore: restore the test for "notify" endpoint
