### 2024-03-26 052ae36
chore: remove unreachable codes
