### 2024-03-26 d431529
[WIP] get NGSI-LD tests to run
Signed-off-by: iripiri <iris.m.koester@gmail.com>
