+++ tests/models/test_ngsi_ld_entity_batch_operation.py
@@ -8,7 +8,7 @@ from filip.clients.ngsi_ld.cb import ContextBrokerLDClient
 from filip.models.ngsi_ld.context import ContextLDEntity, ActionTypeLD
 
 
-class EntitiesBatchOperations(unittest.Testcase):
+class EntitiesBatchOperations(unittest.TestCase):
     """
     Test class for entity endpoints.
     Args:
@@ -98,24 +98,27 @@ class EntitiesBatchOperations(unittest.Testcase):
                                         type=f'filip:object:TypeA') for i in
                           range(0, 10)]
             client.update(entities=entities_a, action_type=ActionTypeLD.CREATE)
-            entity_list = client.get_entity_list()
+            entity_list = client.get_entity_list(entity_type=f'filip:object:TypeA')
+            id_list = [entity.id for entity in entity_list]
+            self.assertEqual(len(entities_a), len(entity_list))
             for entity in entities_a:
-                self.assertIn(entity, entity_list)
-            for entity in entities_a:
+                self.assertIsInstance(entity, ContextLDEntity)
+                self.assertIn(entity.id, id_list)
+            for entity in entity_list:
                 client.delete_entity_by_id(entity_id=entity.id)
         """Test 2"""
         with ContextBrokerLDClient(fiware_header=fiware_header) as client:
-            entities_a = [ContextLDEntity(id=f"urn:ngsi-ld:test:eins",
-                                        type=f'filip:object:TypeA'),
+            entities_b = [ContextLDEntity(id=f"urn:ngsi-ld:test:eins",
+                                        type=f'filip:object:TypeB'),
                           ContextLDEntity(id=f"urn:ngsi-ld:test:eins",
-                                        type=f'filip:object:TypeA')]
+                                        type=f'filip:object:TypeB')]
             try:
-                client.update(entities=entities_a, action_type=ActionTypeLD.CREATE)
-                entity_list = client.get_entity_list()
+                client.update(entities=entities_b, action_type=ActionTypeLD.CREATE)
+                entity_list_b = client.get_entity_list(entity_type=f'filip:object:TypeB')
                 self.assertEqual(len(entity_list), 1)
             except:
                 pass
-            for entity in entities_a:
+            for entity in entity_list_b:
                 client.delete_entity_by_id(entity_id=entity.id)
             
     
@@ -161,20 +164,25 @@ class EntitiesBatchOperations(unittest.Testcase):
                                         type=f'filip:object:TypeUpdate') for i in
                           range(3, 6)]
             client.update(entities=entities_update, action_type=ActionTypeLD.UPDATE)
-            entity_list = client.get_entity_list()
-            for entity in entity_list: 
+            entity_list_a = client.get_entity_list(entity_type=f'filip:object:TypeA')
+            entity_list_b = client.get_entity_list(entity_type=f'filip:object:TypeUpdate')
+            # TODO @lro: does Test 1 still provide any benefit when the entities are retrieved with two calls?
+            for entity in entity_list_a: 
                 if entity.id in ["urn:ngsi-ld:test:0", 
                                  "urn:ngsi-ld:test:1", 
                                  "urn:ngsi-ld:test:2",
                                  "urn:ngsi-ld:test:3"]:
                     
                     self.assertEqual(entity.type, 'filip:object:TypeA')
+            for entity in entity_list_b: 
                 if entity.id in ["urn:ngsi-ld:test:3", 
                                  "urn:ngsi-ld:test:4", 
                                  "urn:ngsi-ld:test:5"]:
                     self.assertEqual(entity.type, 'filip:object:TypeUpdate')
         
-            for entity in entity_list:
+            for entity in entity_list_a:
[…]
