### 2024-03-26 00106b7
[WIP] get NGSI-LD tests to run
Signed-off-by: iripiri <iris.m.koester@gmail.com>
