### 2024-03-27 529a89a
fix: fix parsing and deprecated methods in exercise 6 solution
