### 2024-03-27 eb3702e
Added test function for idPattern
