### 2024-03-27 78e9ad1
fix: fix parsing in exercise 7 solution
