### 2024-03-27 b910989
chore: make opening files consistent with as in other exercises
