### 2024-03-27 f8eac32
Apply suggestions from code review
chore: fix details

Co-authored-by: Marwa Maghnie <68941589+Maghnie@users.noreply.github.com>
