### 2024-03-28 6e61821
Adjusted regex patterns for idPattern test function.
