### 2024-04-02 eb961be
Creating SubWeatherStation entities in OCB
