+++ tutorials/ngsi_v2/e8_multientity_and_expression_language/e8_multientity_and_expression_language_solution.py
@@ -26,6 +26,7 @@ import time
 
 from filip.clients.ngsi_v2 import IoTAClient, ContextBrokerClient
 from filip.models.base import FiwareHeader
+from filip.models.ngsi_v2.context import ContextEntity, NamedContextAttribute
 from filip.models.ngsi_v2.iot import (Device, ServiceGroup, TransportProtocol,
                                       PayloadProtocol, DeviceAttribute,
                                       ExpressionLanguage)
@@ -122,6 +123,17 @@ if __name__ == '__main__':
 
     client.disconnect()
 
+    # Creating SubWeatherStation entities
+    entity1 = ContextEntity(id="urn:ngsi-ld:SubWeatherStation:001",
+                            type="SubWeatherStation")
+    entity1.add_attributes(attrs=[NamedContextAttribute(name="vol", type="Number")])
+    cb_client.post_entity(entity1)
+
+    entity2 = ContextEntity(id="urn:ngsi-ld:SubWeatherStation:002",
+                            type="SubWeatherStation")
+    entity2.add_attributes(attrs=[NamedContextAttribute(name="vol", type="Number")])
+    cb_client.post_entity(entity2)
+
     # TODO: Create a weather station device with multi entity attributes (Number). 'v'
     #  is multiplied by 100 and is a standard attribute. 'v1' and 'v2' are multiplied
     #  by 100 and are multi entity attributes of type SubWeatherStation. The name of
@@ -133,15 +145,15 @@ if __name__ == '__main__':
                      protocol=PayloadProtocol.IOTA_JSON,
                      expressionLanguage=ExpressionLanguage.JEXL,
                      attributes=[DeviceAttribute(object_id="v1", name="vol", type="Number",
-                                                 expression="v1*100",
+                                                 expression="100 * v1",
                                                  entity_name="urn:ngsi-ld:SubWeatherStation:001",
                                                  entity_type="SubWeatherStation"),
                                  DeviceAttribute(object_id="v2", name="vol", type="Number",
-                                                 expression="v2*100",
+                                                 expression="100 * v2",
                                                  entity_name="urn:ngsi-ld:SubWeatherStation:002",
                                                  entity_type="SubWeatherStation"),
                                  DeviceAttribute(object_id="v", name="vol", type="Number",
-                                                 expression="v*100")
+                                                 expression="100 * v")
                                  ]
                      )
     iota_client.post_device(device=device3)
