### 2024-04-02 a5ef21c
doc: add docs for id_pattern
