### 2024-04-03 9c26d2c
Test for ngsi-ld endpoint functions for entity attributes.
