### 2024-04-03 f21b8c0
fix: Session deleted as optional parameter. clients added instead to enable TLS connections for self-signed certificates.
