+++ filip/utils/cleanup.py
@@ -11,13 +11,13 @@ from filip.clients.ngsi_v2 import \
     ContextBrokerClient, \
     IoTAClient, \
     QuantumLeapClient
-from requests import Session
 
 
-def clear_context_broker(url: str, fiware_header: FiwareHeader, session: Session = None ):
+def clear_context_broker(url: str, fiware_header: FiwareHeader, cb_client: ContextBrokerClient = None):
     """
     Function deletes all entities, registrations and subscriptions for a
-    given fiware header
+    given fiware header. To use TLS connection you need to provide the cb_client parameter
+    as an argument with the Session object including the certificate and private key.
 
     Note:
         Always clear the devices first because the IoT-Agent will otherwise
@@ -26,13 +26,18 @@ def clear_context_broker(url: str, fiware_header: FiwareHeader, session: Session
     Args:
         url: Url of the context broker service
         fiware_header: header of the tenant
-        session: session object with set ca.crt and ca.key for TLS support
+        cb_client: enables TLS communication if created with Session object, only needed for self-signed certificates
 
     Returns:
         None
     """
+
     # create client
-    client = ContextBrokerClient(url=url, fiware_header=fiware_header, session=session)
+    if cb_client is None:
+        client = ContextBrokerClient(url=url, fiware_header=fiware_header)
+    else:
+        client = cb_client
+
     # clean entities
     client.delete_entities(entities=client.get_entity_list())
 
@@ -47,21 +52,26 @@ def clear_context_broker(url: str, fiware_header: FiwareHeader, session: Session
     assert len(client.get_registration_list()) == 0
 
 
-def clear_iot_agent(url: Union[str, AnyHttpUrl], fiware_header: FiwareHeader, session: Session = None):
+def clear_iot_agent(url: Union[str, AnyHttpUrl], fiware_header: FiwareHeader, iota_client: IoTAClient = None):
     """
     Function deletes all device groups and devices for a
-    given fiware header
+    given fiware header. To use TLS connection you need to provide the iota_client parameter
+    as an argument with the Session object including the certificate and private key.
 
     Args:
         url: Url of the context broker service
         fiware_header: header of the tenant
-        session: session object with set ca.crt and ca.key for TLS support
+        iota_client: enables TLS communication if created with Session object, only needed for self-signed certificates
 
     Returns:
         None
     """
+
     # create client
-    client = IoTAClient(url=url, fiware_header=fiware_header, session=session)
+    if iota_client is None:
+        client = IoTAClient(url=url, fiware_header=fiware_header)
+    else:
+        client = iota_client
 
     # clear registrations
     for device in client.get_device_list():
@@ -75,13 +85,14 @@ def clear_iot_agent(url: Union[str, AnyHttpUrl], fiware_header: FiwareHeader, se
     assert len(client.get_group_list()) == 0
 
 
-def clear_quantumleap(url: str, fiware_header: FiwareHeader, session: Session = None):
+def clear_quantumleap(url: str, fiware_header: FiwareHeader, ql_client: QuantumLeapClient = None):
     """
-    Function deletes all data for a given fiware header
+    Function deletes all data for a given fiware header. To use TLS connection you need to provide the ql_client parameter
+    as an argument with the Session object including the certificate and private key.
     Args:
         url: Url of the quantumleap service
         fiware_header: header of the tenant
-        session: session object with set ca.crt and ca.key for TLS support
+        ql_client: enables TLS communication if created with Session object, only needed for self-signed certificates
 
     Returns:
         None
@@ -101,7 +112,10 @@ def clear_quantumleap(url: str, fiware_header: FiwareHeader, session: Session = 
         else:
             raise
     # create client
[…]
