### 2024-04-04 602c060
Adding 'timestamp' attribute, which uses the JEXL function 'toisodate'.
