+++ tutorials/ngsi_v2/e8_multientity_and_expression_language/e8_multientity_and_expression_language_solution.py
@@ -23,6 +23,7 @@ recommended language to use is jexl, # which is newer and most powerful.
 """
 # Import packages
 import time
+import datetime
 
 from filip.clients.ngsi_v2 import IoTAClient, ContextBrokerClient
 from filip.models.base import FiwareHeader
@@ -92,7 +93,8 @@ if __name__ == '__main__':
 
     # TODO: Setting expression language to JEXL at Device level with other attributes.
     #  The attribute 'value' (Number) is itself multiplied by 5. The attribute
-    #  'consumption' (String) is the trimmed version of the attribute 'spaces' (String)
+    #  'consumption' (String) is the trimmed version of the attribute 'spaces' (String).
+    #  The attribute 'iso_time' (String) is the current 'timestamp' (Number) transformed into the ISO format.
     device2 = Device(device_id="waste_container_002",
                      entity_name="urn:ngsi-ld:WasteContainer:002",
                      entity_type="WasteContainer",
@@ -103,7 +105,10 @@ if __name__ == '__main__':
                                                  expression="5 * value"),
                                  DeviceAttribute(name="spaces", type="String"),
                                  DeviceAttribute(name="consumption", type="String",
-                                                 expression="spaces | trim")
+                                                 expression="spaces | trim"),
+                                 DeviceAttribute(name="timestamp", type="Number"),
+                                 DeviceAttribute(name="iso_time", type="String",
+                                                 expression="timestamp | toisodate"),
                                  ]
                      )
     iota_client.post_device(device=device2)
@@ -117,9 +122,10 @@ if __name__ == '__main__':
     client.publish(topic=f'/json/{APIKEY}/{device1.device_id}/attrs',
                    payload='{"level": 99, "longitude": 12.0, "latitude": 23.0}')
 
-    # TODO: Publish attributes 'value' and 'spaces' of device2
+    # TODO: Publish attributes 'value', 'spaces' and 'timestamp' (in ms) of device2
     client.publish(topic=f'/json/{APIKEY}/{device2.device_id}/attrs',
-                   payload='{"value": 10, "spaces": "     foobar    "}')
+                   payload=f'{{ "value": 10, "spaces": "     foobar    ",'
+                           f' "timestamp": {datetime.datetime.now().timestamp() * 1000} }}')
 
     client.disconnect()
 
