### 2024-04-04 e4bfc59
chore: format e1_virtual_weatherstation exercise and solution
