+++ tutorials/ngsi_v2/e1_virtual_weatherstation/e1_virtual_weatherstation_solution.py
@@ -3,7 +3,7 @@
 
 # Create a virtual IoT device that simulates the ambient temperature and
 # publishes it via MQTT. The simulation function is already predefined.
-# This exercise to give a simple introduction to the communication via MQTT.
+# This exercise gives a simple introduction to the communication via MQTT.
 
 # The input sections are marked with 'ToDo'
 
@@ -32,14 +32,14 @@ import paho.mqtt.client as mqtt
 from tutorials.ngsi_v2.simulation_model import SimulationModel
 
 # ## Parameters
-# ToDo: Enter your mqtt broker url and port, e.g mqtt://test.mosquitto.org:1883
+# ToDo: Enter your mqtt broker url and port, e.g. mqtt://test.mosquitto.org:1883.
 MQTT_BROKER_URL = "mqtt://test.mosquitto.org:1883"
-# ToDo: If required enter your username and password
+# ToDo: If required, enter your username and password.
 MQTT_USER = ""
 MQTT_PW = ""
 
 # ToDo: Create a unique topic that your weather station will publish on,
-#  e.g. by using a uuid
+#  e.g. by using a uuid.
 UNIQUE_ID = str(uuid4())
 TOPIC_WEATHER_STATION = f"fiware_workshop/{UNIQUE_ID}/weather_station"
 
@@ -49,7 +49,7 @@ TEMPERATURE_MIN = -5  # minimal ambient temperature
 
 T_SIM_START = 0  # simulation start time in seconds
 T_SIM_END = 24 * 60 * 60  # simulation end time in seconds
-COM_STEP = 60 * 60 * 1  # 60 min communication step in seconds
+COM_STEP = 60 * 60  # 60 min communication step in seconds
 
 # ## Main script
 if __name__ == '__main__':
@@ -59,18 +59,18 @@ if __name__ == '__main__':
                                 temp_max=TEMPERATURE_MAX,
                                 temp_min=TEMPERATURE_MIN)
 
-    # define lists to store historical data
+    # define a list for storing historical data
     history_weather_station = []
 
-    # ToDo: create a MQTTv5 client with paho-mqtt
+    # ToDo: Create an MQTTv5 client with paho-mqtt.
     mqttc = mqtt.Client(protocol=mqtt.MQTTv5)
     # set user data if required
     mqttc.username_pw_set(username=MQTT_USER, password=MQTT_PW)
 
     # ToDo: Define a callback function that will be executed when the client
     #  receives message on a subscribed topic. It should decode your message
-    #  and store the information for later in our history
-    #  Note: do not change function's signature!
+    #  and store the information for later in our history.
+    #  Note: Do not change the function's signature!
     def on_message(client, userdata, msg):
         """
         Callback function for incoming messages
@@ -78,7 +78,7 @@ if __name__ == '__main__':
         # decode the payload
         payload = msg.payload.decode('utf-8')
         # ToDo: Parse the payload using the `json` package and write it to
-        #  the history
+        #  the history.
         history_weather_station.append(json.loads(payload))
 
 
@@ -86,7 +86,7 @@ if __name__ == '__main__':
     # or a topic specific callback with `mqttc.message_callback_add()`
     mqttc.on_message = on_message
 
-    # ToDO: connect to the mqtt broker and subscribe to your topic
+    # ToDo: Connect to the mqtt broker and subscribe to your topic.
     mqtt_url = urlparse(MQTT_BROKER_URL)
     mqttc.connect(host=mqtt_url.hostname,
                   port=mqtt_url.port,
@@ -96,27 +96,27 @@ if __name__ == '__main__':
                   clean_start=mqtt.MQTT_CLEAN_START_FIRST_ONLY,
                   properties=None)
 
[…]
