### 2024-04-04 70112a6
chore: format e2_healthcheck exercise and solution
