### 2024-04-04 93e1665
chore: format e3_context_entities exercise and solution
