+++ tutorials/ngsi_v2/e3_context_entities/e3_context_entities.py
@@ -1,11 +1,11 @@
 """
 # # Exercise 3: Context Entities and Relationships
 
-# Create building context entity of type 'Building' according to FIWARE's
+# Create a building context entity of type 'Building' according to FIWARE's
 # SmartData Models with the properties: `id`, `type`, `address`, `category`,
 # https://github.com/smart-data-models/dataModel.Building/blob/master/Building/doc/spec.md
 
-# For the single properties check on the "Data Model description of
+# For the single properties check the "Data Model description of
 # properties" section. The input sections are marked with 'ToDo'
 
 # #### Steps to complete:
@@ -18,12 +18,12 @@
 #    ContextBroker. Afterwards, check if the Context Broker returns the
 #    correct information about your building
 # 6. Create an `opening hours` attribute add them to the server
-# 7. Retrieve the `opening hours` manipulate them and update the model at the
+# 7. Retrieve the `opening hours`, manipulate them and update the model in the
 #    server
 # 8. Repeat the procedure with a thermal zone. Currently, the smart data
 #    models hold no definition of a thermal zone. Therefore, we first only add a
 #    description attribute.
-# 9. Add a `Relationship` attribute to your thermal zone with name
+# 9. Add a `Relationship` attribute to your thermal zone with the name
 #    `refBuilding` and type `Relationship` pointing to your building and post
 #    the model to the context broker
 # 10. Add a `Relationship` attribute to your building with name
@@ -44,16 +44,16 @@ from filip.utils.cleanup import clear_context_broker
 from filip.utils.simple_ql import QueryString
 
 # ## Parameters
-# ToDo: Enter your context broker host and port, e.g http://localhost:1026
+# ToDo: Enter your context broker host and port, e.g. http://localhost:1026.
 CB_URL = "http://localhost:1026"
 
 # ToDo: Change the name of your service to something unique. If you run
-#  on a shared instance this very important in order to avoid user
+#  on a shared instance this is very important in order to avoid user
 #  collisions. You will use this service through the whole tutorial.
-#  If you forget to change it an error will be raised!
+#  If you forget to change it, an error will be raised!
 # FIWARE-Service
 SERVICE = 'filip_tutorial'
-# FIWARE-Servicepath
+# FIWARE-Service path
 SERVICE_PATH = '/'
 
 # ToDo: Path to json-files to store entity data for follow up exercises,
@@ -78,7 +78,7 @@ if __name__ == '__main__':
                                      type="Array",
                                      value=["office"])
 
-    # ToDo: create a property `address` for your building. Follow the full yaml
+    # ToDo: Create a property `address` for your building. Follow the full yaml
     #  description in the specifications. It reuses the specification from
     #  here: https://schema.org/PostalAddress
     address = NamedContextAttribute(name="address",
@@ -89,7 +89,7 @@ if __name__ == '__main__':
 
 
 
-    # ToDo: create a `description` property for your building
+    # ToDo: Create a `description` property for your building.
     building_description = NamedContextAttribute(...)
 
 
@@ -103,10 +103,10 @@ if __name__ == '__main__':
                                    category,
                                    address])
 
-    # ToDo: Create a context broker client and add the fiware_header
+    # ToDo: Create a context broker client and add the fiware_header.
     cbc = ...
     # ToDo: Send your building model to the context broker. Check the client
-    #  for the proper function
+    #  for proper functioning.
     ...
 
     # Update your local building model with the one from the server
@@ -114,9 +114,9 @@ if __name__ == '__main__':
                               entity_type=building.type)
 
     # print your `building model` as json
-    print(f"This is your building model: \n {building.json(indent=2)} \n")
[…]
