### 2024-04-04 b9acf77
chore: format e4_iot_thermal_zone_sensors exercise and solution
