+++ tutorials/ngsi_v2/e4_iot_thermal_zone_sensors/e4_iot_thermal_zone_sensors_solution.py
@@ -2,7 +2,7 @@
 # # Exercise 4: Virtual Thermal Zone
 
 # Create two virtual IoT devices. One of them represents the temperature
-# sensor for the air temperature of a the thermal zone, whereas the second
+# sensor for the air temperature of a thermal zone, whereas the second
 # represents a virtual weather station. Both devices publish their values to
 # the platform via MQTT. Use the simulation model of
 # e1_virtual_weatherstation.py
@@ -40,35 +40,33 @@ from filip.utils.cleanup import clear_context_broker, clear_iot_agent
 from tutorials.ngsi_v2.simulation_model import SimulationModel
 
 # ## Parameters
-# ToDo: Enter your context broker host and port, e.g. http://localhost:1026
+# ToDo: Enter your context broker host and port, e.g. http://localhost:1026.
 CB_URL = "http://localhost:1026"
-# ToDo: Enter your IoT-Agent host and port, e.g. http://localhost:4041
+# ToDo: Enter your IoT-Agent host and port, e.g. http://localhost:4041.
 IOTA_URL = "http://localhost:4041"
-# ToDo: Enter your mqtt broker url, e.g mqtt://test.mosquitto.org:1883
+# ToDo: Enter your mqtt broker url, e.g. mqtt://test.mosquitto.org:1883.
 MQTT_BROKER_URL = "mqtt://localhost:1883"
-# ToDo: If required, enter your username and password
+# ToDo: If required, enter your username and password.
 MQTT_USER = ""
 MQTT_PW = ""
 
 # ToDo: Change the name of your service to something unique. If you run
-#  on a shared instance this very important in order to avoid user
+#  on a shared instance this is very important in order to avoid user
 #  collisions. You will use this service through the whole tutorial.
-#  If you forget to change it an error will be raised!
+#  If you forget to change it, an error will be raised!
 # FIWARE-Service
 SERVICE = 'filip_tutorial'
-# FIWARE-Servicepath
+# FIWARE-Service path
 SERVICE_PATH = '/'
 
-# ToDo: Change the APIKEY to something unique. This represent the "token"
-#  for IoT devices to connect (send/receive data ) with the platform. In the
+# ToDo: Change the APIKEY to something unique. This represents the "token"
+#  for IoT devices to connect (send/receive data) with the platform. In the
 #  context of MQTT, APIKEY is linked with the topic used for communication.
 APIKEY = 'your_apikey'
 
-# Path to json-files to device configuration data for follow-up exercises
-WRITE_GROUPS_FILEPATH = Path(
-    "../e4_iot_thermal_zone_sensors_solution_groups.json")
-WRITE_DEVICES_FILEPATH = Path(
-    "../e4_iot_thermal_zone_sensors_solution_devices.json")
+# path to json-files to device configuration data for follow-up exercises
+WRITE_GROUPS_FILEPATH = Path("../e4_iot_thermal_zone_sensors_solution_groups.json")
+WRITE_DEVICES_FILEPATH = Path("../e4_iot_thermal_zone_sensors_solution_devices.json")
 
 # set parameters for the temperature simulation
 TEMPERATURE_MAX = 10  # maximal ambient temperature
@@ -99,15 +97,15 @@ if __name__ == '__main__':
     history_weather_station = []
     history_zone_temperature_sensor = []
 
-    # create a service group and add it to your
+    # create a service group with your api key
     service_group = ServiceGroup(apikey=APIKEY,
                                  resource="/iot/json")
 
-    # ToDo: create two IoTA-MQTT devices for the weather station and the zone
+    # ToDo: Create two IoTA-MQTT devices for the weather station and the zone
     #  temperature sensor. Also add the simulation time as `active attribute`
     #  to each device!
     # create the weather station device
[…]
