+++ tutorials/ngsi_v2/e4_iot_thermal_zone_sensors/e4_iot_thermal_zone_sensors.py
@@ -1,9 +1,9 @@
 """
 # # Exercise 4: Virtual Thermal Zone
 
-# Create two virtual IoT device. One of them represents the temperature
-# sensor for the air temperature of a the thermal zone, whereas the second
-# represents a virtual weather station. Both devices publish there values to
+# Create two virtual IoT devices. One of them represents the temperature
+# sensor for the air temperature of a thermal zone, whereas the second
+# represents a virtual weather station. Both devices publish their values to
 # the platform via MQTT. Use the simulation model of
 # e1_virtual_weatherstation.py
 #
@@ -11,7 +11,7 @@
 #
 # #### Steps to complete:
 # 1. Set up the missing parameters in the parameter section
-# 2. Create a service group and two devices
+# 2. Create a service group and two corresponding devices
 # 3. Provision the service group and the devices
 # 4. Create an MQTT client using the filip.client.mqtt package and register
 #    your service group and your devices
@@ -31,7 +31,7 @@ import matplotlib.pyplot as plt
 import paho.mqtt.client as mqtt
 
 # import from filip
-from filip.clients.ngsi_v2 import ContextBrokerClient
+from filip.clients.ngsi_v2 import ContextBrokerClient, IoTAClient
 from filip.clients.mqtt import IoTAMQTTClient
 from filip.models.base import FiwareHeader
 from filip.models.ngsi_v2.iot import Device, DeviceAttribute, ServiceGroup
@@ -40,30 +40,31 @@ from filip.utils.cleanup import clear_context_broker, clear_iot_agent
 from tutorials.ngsi_v2.simulation_model import SimulationModel
 
 # ## Parameters
-# ToDo: Enter your context broker host and port, e.g. http://localhost:1026
+# ToDo: Enter your context broker host and port, e.g. http://localhost:1026.
 CB_URL = "http://localhost:1026"
-# ToDo: Enter your IoT-Agent host and port, e.g. http://localhost:4041
+# ToDo: Enter your IoT-Agent host and port, e.g. http://localhost:4041.
 IOTA_URL = "http://localhost:4041"
-# ToDo: Enter your mqtt broker url, e.g mqtt://test.mosquitto.org:1883
+# ToDo: Enter your mqtt broker url, e.g. mqtt://test.mosquitto.org:1883.
 MQTT_BROKER_URL = "mqtt://localhost:1883"
-# ToDo: If required enter your username and password
+# ToDo: If required, enter your username and password.
 MQTT_USER = ""
-MQTT_PW =  ""
+MQTT_PW = ""
+
 # ToDo: Change the name of your service to something unique. If you run
-#  on a shared instance this very important in order to avoid user
+#  on a shared instance this is very important in order to avoid user
 #  collisions. You will use this service through the whole tutorial.
-#  If you forget to change it an error will be raised!
+#  If you forget to change it, an error will be raised!
 # FIWARE-Service
 SERVICE = 'filip_tutorial'
-# FIWARE-Servicepath
+# FIWARE-Service path
 SERVICE_PATH = '/'
 
-# ToDo: Change the APIKEY to something unique. This represent the "token"
-#  for IoT devices to connect (send/receive data ) with the platform. In the
+# ToDo: Change the APIKEY to something unique. This represents the "token"
+#  for IoT devices to connect (send/receive data) with the platform. In the
 #  context of MQTT, APIKEY is linked with the topic used for communication.
 APIKEY = 'your_apikey'
 
-# Path to json-files to device configuration data for follow up exercises
+# path to json-files to device configuration data for follow-up exercises
[…]
