### 2024-04-04 5bf544a
chore:
Define ToDos for the further validation of nested propertys
