### 2024-04-05 0d4d85c
Added test to reproduce cleanup function order error.
