### 2024-04-09 cb632eb
run NGSI-LD batch tests, revise implementation
Signed-off-by: iripiri <iris.m.koester@gmail.com>
