### 2024-04-09 baab79d
run NGSI-LD batch tests, revise implementation
Signed-off-by: iripiri <iris.m.koester@gmail.com>
