+++ tests/models/test_ngsi_ld_entity_batch_operation.py
@@ -65,6 +65,19 @@ class EntitiesBatchOperations(unittest.TestCase):
     #     if 1 == 1:
     #         self.assertNotEqual(1,2)
     #     pass 
+
+    def tearDown(self) -> None:
+        """
+        Cleanup entities from test server
+        """
+        entity_test_types = ["filip:object:TypeA", "filip:object:TypeB", "filip:object:TypeUpdate", "filip:object:TypeDELETE"]
+
+        fiware_header = FiwareLDHeader()
+        with ContextBrokerLDClient(fiware_header=fiware_header) as client:
+            for entity_type in entity_test_types:
+                entity_list = client.get_entity_list(entity_type=entity_type)
+                for entity in entity_list:
+                    client.delete_entity_by_id(entity_id=entity.id)
     
     def test_entity_batch_operations_create(self) -> None:
         """
@@ -112,14 +125,16 @@ class EntitiesBatchOperations(unittest.TestCase):
                                         type=f'filip:object:TypeB'),
                           ContextLDEntity(id=f"urn:ngsi-ld:test:eins",
                                         type=f'filip:object:TypeB')]
+            entity_list_b = []
             try:
                 client.update(entities=entities_b, action_type=ActionTypeLD.CREATE)
                 entity_list_b = client.get_entity_list(entity_type=f'filip:object:TypeB')
                 self.assertEqual(len(entity_list), 1)
             except:
                 pass
-            for entity in entity_list_b:
-                client.delete_entity_by_id(entity_id=entity.id)
+            finally:
+                for entity in entity_list_b:
+                    client.delete_entity_by_id(entity_id=entity.id)
             
     
     def test_entity_operations_update(self) -> None:
@@ -216,9 +231,8 @@ class EntitiesBatchOperations(unittest.TestCase):
                 client.delete_entity_by_id(entity_id=entity.id)   
                 
     # TODO @lro: 
-    # - using curl commands, upsert replace does not work while changing the type
-    # seems like only attributes can be replaced
-    # - a test with empty array would and/or containing null value also be good,
+    # - changing the entity type needs to be tested with new release, did not work so far
+    # - a test with empty array and/or containing null value would also be good,
     # should result in BadRequestData error
     def test_entity_operations_upsert(self) -> None:
         """
@@ -262,6 +276,7 @@ class EntitiesBatchOperations(unittest.TestCase):
             entities_upsert = [ContextLDEntity(id=f"urn:ngsi-ld:test:{str(i)}",
                                         type=f'filip:object:TypeUpdate') for i in
                           range(2, 6)]
+            # TODO: this should work with newer release of orion-ld broker
             client.update(entities=entities_upsert, action_type=ActionTypeLD.UPSERT, update_format="update")
 
             # read entities from broker and check that entities were not replaced
@@ -327,7 +342,7 @@ class EntitiesBatchOperations(unittest.TestCase):
                 client.delete_entity_by_id(entity_id=entity.id)  
 
 
-    def aatest_entity_operations_delete(self) -> None:
+    def test_entity_operations_delete(self) -> None:
         """
         Batch entity delete.
         Args:
@@ -357,7 +372,7 @@ class EntitiesBatchOperations(unittest.TestCase):
         fiware_header = FiwareLDHeader()
         with ContextBrokerLDClient(fiware_header=fiware_header) as client:
             entities_delete = [ContextLDEntity(id=f"urn:ngsi-ld:test:{str(i)}",
-                            type=f'filip:object:TypeA') for i in
+                            type=f'filip:object:TypeDELETE') for i in
                 range(0, 1)]
             with self.assertRaises(Exception):
[…]
