### 2024-04-09 65e405e
feat: add validator for temporal query
