+++ filip/models/ngsi_ld/subscriptions.py
@@ -1,6 +1,7 @@
-from typing import List, Optional, Union
-from pydantic import ConfigDict, BaseModel, Field, HttpUrl, AnyUrl,\
-    field_validator
+from typing import List, Optional, Union, Literal
+from pydantic import ConfigDict, BaseModel, Field, HttpUrl, AnyUrl, \
+    field_validator, model_validator
+import dateutil.parser
 
 
 class EntityInfo(BaseModel):
@@ -154,23 +155,68 @@ class NotificationParams(BaseModel):
 
 
 class TemporalQuery(BaseModel):
-    timerel: str = Field(
+    """
+    Temporal query according to NGSI-LD Spec section 5.2.21
+
+    timerel:
+        Temporal relationship, one of "before", "after" and "between".
+        "before": before the time specified by timeAt.
+        "after": after the time specified by timeAt.
+        "between": after the time specified by timeAt and before the time specified by
+            endtimeAt
+    timeAt:
+        A DateTime object following ISO 8061, e.g. 2007-12-24T18:21Z
+    endTimeAt (optional):
+        A DateTime object following ISO 8061, e.g. 2007-12-24T18:21Z
+        Only required when timerel="between"
+    timeproperty: str
+        Representing a Propertyname of the Property that contains the temporal data that
+        will be used to resolve the temporal query. If not specified, the default is
+        "observedAt"
+
+    """
+    model_config = ConfigDict(populate_by_name=True)
+    timerel: Literal['before', 'after', 'between'] = Field(
         ...,
-        description="String representing the temporal relationship as defined by clause 4.11 (Allowed values: 'before', 'after', and 'between')"
+        description="String representing the temporal relationship as defined by clause "
+                    "4.11 (Allowed values: 'before', 'after', and 'between') "
     )
     timeAt: str = Field(
         ...,
-        description="String representing the timeAt parameter as defined by clause 4.11. It shall be a DateTime"
+        description="String representing the timeAt parameter as defined by clause "
+                    "4.11. It shall be a DateTime "
     )
     endTimeAt: Optional[str] = Field(
         default=None,
-        description="String representing the endTimeAt parameter as defined by clause 4.11. It shall be a DateTime. Cardinality shall be 1 if timerel is equal to 'between'"
+        description="String representing the endTimeAt parameter as defined by clause "
+                    "4.11. It shall be a DateTime. Cardinality shall be 1 if timerel is "
+                    "equal to 'between' "
     )
     timeproperty: Optional[str] = Field(
         default=None,
-        description="String representing a Property name. The name of the Property that contains the temporal data that will be used to resolve the temporal query. If not specified,"
+        description="String representing a Property name. The name of the Property that "
+                    "contains the temporal data that will be used to resolve the "
+                    "temporal query. If not specified, "
     )
-    model_config = ConfigDict(populate_by_name=True)
+
+    @field_validator("timeAt", "endTimeAt")
+    @classmethod
+    def check_uri(cls, v: str):
+        if not v:
+            return v
+        else:
+            try:
+                dateutil.parser.isoparse(v)
+            except ValueError:
+                raise ValueError("timeAt must be in ISO8061 format")
+            return v
+
+    # when timerel=between, endTimeAt must be specified
+    @model_validator(mode='after')
+    def check_passwords_match(self) -> 'TemporalQuery':
+        if self.timerel == "between" and self.endTimeAt is None:
+            raise ValueError('When timerel="between", endTimeAt must be specified')
+        return self
 
 
 class Subscription(BaseModel):
