### 2024-04-09 8e59226
feat: test for temporal query
