+++ tests/models/test_ngsi_ld_subscriptions.py
@@ -8,9 +8,9 @@ from pydantic import ValidationError
 # from filip.clients.ngsi_v2 import ContextBrokerClient
 from filip.models.ngsi_ld.subscriptions import \
     Subscription, \
-    Endpoint, NotificationParams, EntityInfo
+    Endpoint, NotificationParams, EntityInfo, TemporalQuery
 from filip.models.base import FiwareHeader
-from filip.utils.cleanup import clear_all, clean_test
+from filip.utils.cleanup import clear_all
 from tests.config import settings
 
 
@@ -131,7 +131,23 @@ class TestLDSubscriptions(unittest.TestCase):
         Returns:
 
         """
-        pass
+        example0_temporalQ = {
+            "timerel": "before",
+            "timeAt": "2017-12-13T14:20:00Z"
+        }
+        self.assertEqual(example0_temporalQ,
+                         TemporalQuery.model_validate(example0_temporalQ).model_dump(
+                             exclude_unset=True)
+                         )
+
+        example1_temporalQ = {
+            "timerel": "after",
+            "timeAt": "2017-12-13T14:20:00Z"
+        }
+        self.assertEqual(example1_temporalQ,
+                         TemporalQuery.model_validate(example1_temporalQ).model_dump(
+                             exclude_unset=True)
+                         )
 
         example2_temporalQ = {
             "timerel": "between",
@@ -140,7 +156,9 @@ class TestLDSubscriptions(unittest.TestCase):
             "timeproperty": "modifiedAt"
         }
         self.assertEqual(example2_temporalQ,
-                         TemporalQuery.model_validate(example2_temporalQ).model_dump())
+                         TemporalQuery.model_validate(example2_temporalQ).model_dump(
+                             exclude_unset=True)
+                         )
 
         example3_temporalQ = {
             "timerel": "between",
@@ -159,7 +177,7 @@ class TestLDSubscriptions(unittest.TestCase):
         example5_temporalQ = {
             "timerel": "between",
             "timeAt": "2017-12-13T14:20:00Z",
-            "endTimeAt": "2017-12-13T14:40:00Z",
+            "endTimeAt": "14:40:00Z",
             "timeproperty": "modifiedAt"
         }
         with self.assertRaises(ValueError):
