### 2024-04-09 00ea4e0
chore: format e5_iot_thermal_zone_control exercise and solution
