+++ tutorials/ngsi_v2/e5_iot_thermal_zone_control/e5_iot_thermal_zone_control_solution.py
@@ -3,7 +3,7 @@
 
 # Create a virtual IoT device that simulates a heater for your
 # thermal zone. The heater can be turned on and off via a simple hysteresis
-# controller. The devices from e4_iot_thermal_zone_sensors.py will loaded
+# controller. The devices from e4_iot_thermal_zone_sensors.py will be loaded
 # from the stored *.json-files.
 
 # The input sections are marked with 'ToDo'
@@ -13,18 +13,18 @@
 # 2. Retrieve the service group and device configurations of already existing
 #    devices from the IoT-Agent
 # 3. Create a third device configuration for a heater holding a command
-#    for turning it `on` and `off`device and post it to the server
+#    for turning it `on` and `off` and post it to the server
 # 4. Create an MQTT client using the filip.client.mqtt package and register
 #    your service group and your devices
 # 4. Define a callback function that will be executed when the client
-#    receives a command. Decode the message and set the update the state in
+#    receives a command. Decode the message and set the update state in
 #    simulation model. Afterwards, acknowledge the command using the api of the
 #    IoTAMQTTClient.
 # 5. Add the callback for your heater device to the IoTAMQTTClient
 # 6. Create an MQTT subscription for asynchronous communication that
 #    gets triggered when the temperature attribute changes.
 # 7. Write a second callback that represents your controller. It should get
-#    triggered when the MQTTClient receive a notification message due to your
+#    triggered when the MQTTClient receives a notification message due to your
 #    subscription. Add the callback to your MQTTClient using the original
 #    paho-api (`message_callback_add`)
 # 8. Run the simulation and plot
@@ -58,48 +58,48 @@ from filip.utils.cleanup import clear_context_broker, clear_iot_agent
 from tutorials.ngsi_v2.simulation_model import SimulationModel
 
 # ## Parameters
-# ToDo: Enter your context broker host and port, e.g http://localhost:1026
+# ToDo: Enter your context broker host and port, e.g http://localhost:1026.
 CB_URL = "http://localhost:1026"
-# ToDo: Enter your IoT-Agent host and port, e.g http://localhost:4041
+# ToDo: Enter your IoT-Agent host and port, e.g http://localhost:4041.
 IOTA_URL = "http://localhost:4041"
-# ToDo: Enter your mqtt broker url, e.g mqtt://test.mosquitto.org:1883
+# ToDo: Enter your mqtt broker url, e.g mqtt://test.mosquitto.org:1883.
 MQTT_BROKER_URL_EXPOSED = "mqtt://localhost:1883"
-# ToDo: Enter your mqtt broker url, e.g mqtt://mosquitto:1883
+# ToDo: Enter your mqtt broker url, e.g mqtt://mosquitto:1883.
 MQTT_BROKER_URL_INTERNAL = "mqtt://mosquitto:1883"
-# ToDo: If required enter your username and password
+# ToDo: If required, enter your username and password.
 MQTT_USER = ""
 MQTT_PW = ""
 
 # ToDo: Change the name of your service to something unique. If you run
-#  on a shared instance this very important in order to avoid user
+#  on a shared instance this is very important in order to avoid user
 #  collisions. You will use this service through the whole tutorial.
-#  If you forget to change it an error will be raised!
+#  If you forget to change it, an error will be raised!
 # FIWARE-Service
 SERVICE = 'filip_tutorial'
-# FIWARE-Servicepath
+# FIWARE-Service path
 SERVICE_PATH = '/'
 
-# ToDo: Change the APIKEY to something unique. This represent the "token"
-#  for IoT devices to connect (send/receive data ) with the platform. In the
+# ToDo: Change the APIKEY to something unique. This represents the "token"
+#  for IoT devices to connect (send/receive data) with the platform. In the
 #  context of MQTT, APIKEY is linked with the topic used for communication.
[…]
