### 2024-04-10 8416c70
chore: format e6_timeseries_data exercise and solution
