### 2024-04-10 17901f6
chore: format e7_semantic_iot exercise and solution
