+++ tutorials/ngsi_v2/e7_semantic_iot/e7_semantic_iot.py
@@ -2,7 +2,7 @@
 # # Exercise 7: Semantic IoT Systems
 #
 # We now want to add a semantic meaning to our measurements. Therefore we
-# semantically connect to the context entities that we created in
+# semantically connect the context entities that we created in
 # `e3_context_entities.py`
 
 # The input sections are marked with 'ToDo'
@@ -14,14 +14,15 @@
 # 3. Add relationships that connect the thermal zone temperature sensor and
 #    the heater to the building, and vice versa
 # 4. Retrieve all entities and print them
-# 5. Congratulations you are now ready to build your own semantic systems for
+# 5. Congratulations! You are now ready to build your own semantic systems. For
 #    advanced semantic functions check on our semantics examples
 """
 
 # ## Import packages
+import json
 from pathlib import Path
 from typing import List
-from pydantic import parse_file_as
+from pydantic import TypeAdapter
 
 # import from filip
 from filip.clients.ngsi_v2 import \
@@ -42,32 +43,40 @@ from filip.utils.cleanup import \
     clear_iot_agent
 
 # ## Parameters
-# ToDo: Enter your context broker host and port, e.g http://localhost:1026
+# ToDo: Enter your context broker host and port, e.g. http://localhost:1026.
 CB_URL = "http://localhost:1026"
-# ToDo: Enter your IoT-Agent host and port, e.g http://localhost:4041
+# ToDo: Enter your IoT-Agent host and port, e.g. http://localhost:4041.
 IOTA_URL = "http://localhost:4041"
 
 # FIWARE-Service
 SERVICE = 'filip_tutorial'
-# FIWARE-Servicepath
+# FIWARE-Service path
 # ToDo: Change the name of your service-path to something unique. If you run
-#  on a shared instance this very important in order to avoid user
+#  on a shared instance this is very important in order to avoid user
 #  collisions. You will use this service path through the whole tutorial.
-#  If you forget to change it an error will be raised!
+#  If you forget to change it, an error will be raised!
 SERVICE_PATH = '/'
 
-# ToDo: Change the APIKEY to something unique. This represent the "token"
-#  for IoT devices to connect (send/receive data ) with the platform. In the
+# ToDo: Change the APIKEY to something unique. This represents the "token"
+#  for IoT devices to connect (send/receive data) with the platform. In the
 #  context of MQTT, APIKEY is linked with the topic used for communication.
 APIKEY = 'your_apikey'
 
-# Path to read json-files from previous exercises
+# path to read json-files from previous exercises
 READ_GROUPS_FILEPATH = \
-    Path("../e5_iot_thermal_zone_control_groups.json")
+    Path("../e5_iot_thermal_zone_control_solution_groups.json")
 READ_DEVICES_FILEPATH = \
-    Path("../e5_iot_thermal_zone_control_devices.json")
+    Path("../e5_iot_thermal_zone_control_solution_devices.json")
 READ_ENTITIES_FILEPATH = \
-    Path("../e3_context_entities_entities.json")
+    Path("../e3_context_entities_solution_entities.json")
+
+# opening the files
+with (open(READ_GROUPS_FILEPATH, 'r') as groups_file,
+      open(READ_DEVICES_FILEPATH, 'r') as devices_file,
+      open(READ_ENTITIES_FILEPATH, 'r') as entities_file):
+    json_groups = json.load(groups_file)
+    json_devices = json.load(devices_file)
+    json_entities = json.load(entities_file)
 
 # ## Main script
 if __name__ == '__main__':
@@ -78,10 +87,10 @@ if __name__ == '__main__':
     clear_iot_agent(url=IOTA_URL, fiware_header=fiware_header)
     clear_context_broker(url=CB_URL, fiware_header=fiware_header)
 
-    # Create clients and restore devices and groups from file
[…]
