### 2024-04-10 5fe9468
format: fix typo in exercise 1
