### 2024-04-10 0a610a5
fix: fix plotting in exercise 5
