### 2024-04-10 8f98032
Test for NGSI-LD endpoint subscription (get all subscriptions).
