+++ tests/models/test_ngsi_ld_subscription.py
@@ -5,15 +5,19 @@ import json
 import unittest
 
 from pydantic import ValidationError
-from filip.clients.ngsi_v2 import ContextBrokerClient
-from filip.models.ngsi_v2.subscriptions import \
-    Mqtt, \
-    MqttCustom, \
+
+from filip.clients.ngsi_ld.cb import ContextBrokerLDClient
+from filip.models.base import FiwareLDHeader
+from filip.models.ngsi_ld.context import \
+    ContextProperty, \
+    NamedContextProperty
+from filip.models.ngsi_ld.subscriptions import \
+    Endpoint, \
+    NotificationParams, \
     Subscription
-# MQtt should be the same just the sub has to be changed to fit LD
-from filip.models.base import FiwareHeader
 from filip.utils.cleanup import clear_all, clean_test
 from tests.config import settings
+from random import randint
 
 class TestSubscriptions(unittest.TestCase):
     """
@@ -26,11 +30,11 @@ class TestSubscriptions(unittest.TestCase):
         Returns:
             None
         """
-        self.fiware_header = FiwareHeader(
+        self.fiware_header = FiwareLDHeader(
             service=settings.FIWARE_SERVICE,
             service_path=settings.FIWARE_SERVICEPATH)
-        self.mqtt_url = "mqtt://test.de:1883"
-        self.mqtt_topic = '/filip/testing'
+        # self.mqtt_url = "mqtt://test.de:1883"
+        # self.mqtt_topic = '/filip/testing'
         # self.notification =  {
         #     "attributes": ["filling", "controlledAsset"],
         #     "format": "keyValues",
@@ -39,244 +43,144 @@ class TestSubscriptions(unittest.TestCase):
         #     "accept": "application/json"
         #     }
         # }
-        self.sub_dict = {
-            "description": "One subscription to rule them all",
-            "type": "Subscription",
-            "entities": [
-                {
-                    "type": "FillingLevelSensor",
-                }
-            ],
-            "watchedAttributes": ["filling"],
-            "q": "filling>0.6",
-            "notification": {
-                "attributes": ["filling", "controlledAsset"],
-                "format": "keyValues",
-                "endpoint": {
-                "uri": "http://test:1234/subscription/low-stock-farm001-ngsild",
-                "accept": "application/json"
-                }
-            },
-            "@context": "http://context/ngsi-context.jsonld"
-        }
+        self.cb_client = ContextBrokerLDClient()
+        self.endpoint_http = Endpoint(**{
+            "uri": "http://my.endpoint.org/notify",
+            "accept": "application/json"
+        })
 
-    # def test_notification_models(self):
-    #     """
-    #     Test notification models
-    #     """
-    #     # Test url field sub field validation
-    #     with self.assertRaises(ValidationError):
-    #         Mqtt(url="brokenScheme://test.de:1883",
-    #              topic='/testing')
-    #     with self.assertRaises(ValidationError):
-    #         Mqtt(url="mqtt://test.de:1883",
-    #              topic='/,t')
-    #     mqtt = Mqtt(url=self.mqtt_url,
-    #                 topic=self.mqtt_topic)
-    #     mqttCustom = MqttCustom(url=self.mqtt_url,
-    #                             topic=self.mqtt_topic)
-
-    #     # Test validator for conflicting fields
-    #     notification = Notification.model_validate(self.notification)
-    #     with self.assertRaises(ValidationError):
-    #         notification.mqtt = mqtt
-    #     with self.assertRaises(ValidationError):
-    #         notification.mqtt = mqttCustom
-
-    #     # test onlyChangedAttrs-field
-    #     notification = Notification.model_validate(self.notification)
[…]
