### 2024-04-10 e45350b
Progress enpoint test for subscription (delete subscription).
