### 2024-04-11 05bbb4d
added teardown, updated get/post/delete tests and corresponding implementation
Signed-off-by: iripiri <iris.m.koester@gmail.com>
