### 2024-04-11 a53e39f
added teardown, updated get/post/delete tests and corresponding implementation
Signed-off-by: iripiri <iris.m.koester@gmail.com>
