+++ filip/clients/ngsi_ld/cb.py
@@ -237,9 +237,70 @@ class ContextBrokerLDClient(BaseHttpClient):
 
     GeometryShape = Literal["Point", "MultiPoint", "LineString", "MultiLineString", "Polygon", "MultiPolygon"]
 
+    def get_entity(self,
+                   entity_id: str,
+                   entity_type: str = None,
+                   attrs: List[str] = None,
+                   response_format: Union[AttrsFormat, str] =
+                   AttrsFormat.NORMALIZED,
+                   **kwargs  # TODO how to handle metadata?
+                   ) \
+            -> Union[ContextLDEntity, ContextLDEntityKeyValues, Dict[str, Any]]:
+        """
+        This operation must return one entity element only, but there may be
+        more than one entity with the same ID (e.g. entities with same ID but
+        different types). In such case, an error message is returned, with
+        the HTTP status code set to 409 Conflict.
+
+        Args:
+            entity_id (String): Id of the entity to be retrieved
+            entity_type (String): Entity type, to avoid ambiguity in case
+                there are several entities with the same entity id.
+            attrs (List of Strings): List of attribute names whose data must be
+                included in the response. The attributes are retrieved in the
+                order specified by this parameter.
+                See "Filtering out attributes and metadata" section for more
+                detail. If this parameter is not included, the attributes are
+                retrieved in arbitrary order, and all the attributes of the
+                entity are included in the response.
+                Example: temperature, humidity.
+            response_format (AttrsFormat, str): Representation format of
+                response
+        Returns:
+            ContextEntity
+        """
+        url = urljoin(self.base_url, f'{self._url_version}/entities/{entity_id}')
+        headers = self.headers.copy()
+        params = {}
+        if entity_type:
+            params.update({'type': entity_type})
+        if attrs:
+            params.update({'attrs': ','.join(attrs)})
+
+        if response_format:
+            if response_format not in list(AttrsFormat):
+                raise ValueError(f'Value must be in {list(AttrsFormat)}')
+            #params.update({'options': response_format})
+
+        try:
+            res = self.get(url=url, params=params, headers=headers)
+            if res.ok:
+                self.logger.info("Entity successfully retrieved!")
+                self.logger.debug("Received: %s", res.json())
+                if response_format == AttrsFormat.NORMALIZED:
+                    return ContextLDEntity(**res.json())
+                if response_format == AttrsFormat.KEY_VALUES:
+                    return ContextLDEntityKeyValues(**res.json())
+                return res.json()
+            res.raise_for_status()
+        except requests.RequestException as err:
+            msg = f"Could not load entity {entity_id}"
+            self.log_error(err=err, msg=msg)
+            raise
+
     def get_entity_list(self,
                         entity_id: Optional[str] = None,
-                        id_pattern: Optional[str] = None,
+                        id_pattern: Optional[str] = ".*",
                         entity_type: Optional[str] = None,
                         attrs: Optional[List[str]] = None,
                         q: Optional[str] = None,
@@ -248,7 +309,7 @@ class ContextBrokerLDClient(BaseHttpClient):
                         coordinates: Optional[str] = None,
                         geoproperty: Optional[str] = None,
                         csf: Optional[str] = None,
-                        limit: Optional[PositiveInt] = None,
+                        limit: Optional[PositiveInt] = 100,
                         response_format: Optional[Union[AttrsFormat, str]] = AttrsFormat.NORMALIZED.value,
                         ) -> Union[Dict[str, Any]]:
 
@@ -404,13 +465,13 @@ class ContextBrokerLDClient(BaseHttpClient):
 
     def delete_entity_by_id(self,
[…]
