+++ tests/models/test_ngsi_ld_entities.py
@@ -18,7 +18,8 @@ from filip.models.ngsi_ld.context import \
     ContextLDEntity, \
     ContextProperty, \
     ContextRelationship, \
-    NamedContextProperty
+    NamedContextProperty, \
+    ActionTypeLD
 import requests
 
 class TestEntities(unittest.TestCase):
@@ -40,7 +41,8 @@ class TestEntities(unittest.TestCase):
         self.mqtt_url = "mqtt://test.de:1883"
         self.mqtt_topic = '/filip/testing'
 
-        CB_URL = "http://localhost:1026"
+        #CB_URL = "http://localhost:1026"
+        CB_URL = "http://137.226.248.200:1027"
         self.cb_client = ContextBrokerLDClient(url=CB_URL,
                                     fiware_header=self.fiware_header)
         
@@ -54,9 +56,20 @@ class TestEntities(unittest.TestCase):
                                     #   type="room",
                                     #   data={})
         self.entity_2 = ContextLDEntity(id="urn:ngsi-ld:room2",
-                                        type="room",
-                                      data={})
-    
+                                        type="room")
+
+    def tearDown(self) -> None:
+        """
+        Cleanup entities from test server
+        """
+        entity_test_types = ["MyType", "room"]
+
+        fiware_header = FiwareLDHeader()
+        with ContextBrokerLDClient(fiware_header=fiware_header) as client:
+            for entity_type in entity_test_types:
+                entity_list = client.get_entity_list(entity_type=entity_type)
+                for entity in entity_list:
+                    client.delete_entity_by_id(entity_id=entity.id)
 
     def test_get_entites(self):
         """
@@ -110,16 +123,16 @@ class TestEntities(unittest.TestCase):
             Post enitity with entity_ID and entity_name
             if return != 201:
                 Raise Error
-            Get enitity list 
+            Get entity list
             If entity with entity_ID is not on entity list:
                 Raise Error
         Test 2: 
-            Post enitity with entity_ID and entity_name
+            Post entity with entity_ID and entity_name
             Post entity with the same entity_ID and entity_name as before
             If return != 409: 
                 Raise Error
-            Get enitity list
-            If there are duplicates on enity list:
+            Get entity list
+            If there are duplicates on entity list:
                 Raise Error
         Test 3: 
             Post an entity with an entity_ID and without an entity_name 
@@ -132,18 +145,22 @@ class TestEntities(unittest.TestCase):
             post two entities with the same enitity id but different entity type-> should throw error.
         """
         """Test1"""
-        ret_post = self.cb_client.post_entity(entity=self.entity)
-        # Raise already done in cb
-        entity_list = self.cb_client.get_entity_list()
-        self.assertIn(self.entity, entity_list)   
+        self.cb_client.post_entity(entity=self.entity)
+        entity_list = self.cb_client.get_entity_list(entity_type=self.entity.type)
+        self.assertEqual(len(entity_list), 1)
+        self.assertEqual(entity_list[0].id, self.entity.id)
+        self.assertEqual(entity_list[0].type, self.entity.type)
+        self.assertEqual(entity_list[0].testtemperature["value"], self.entity.testtemperature["value"])
         
         """Test2"""
         self.entity_identical= self.entity.model_copy()
-        ret_post = self.cb_client.post_entity(entity=self.entity_identical)
-        # What is gonna be the return? Is already an error being raised?
-        entity_list = self.cb_client.get_entity_list()
-        for element in entity_list: 
-            self.assertNotEqual(element.id, self.entity.id)
+        with self.assertRaises(requests.exceptions.HTTPError) as contextmanager:
[…]
