### 2024-04-15 92f2580
Adding validation for JEXL expression language of Device attributes and setting default expression language to JEXL.
