### 2024-04-16 3bd8e2f
fix: fix opening files for <python 3.10
