+++ tutorials/ngsi_v2/e3_context_entities/e3_context_entities_solution.py
@@ -57,7 +57,10 @@ SERVICE = 'filip_tutorial'
 SERVICE_PATH = '/'
 
 # ToDo: Path to json-files to store entity data for follow up exercises,
-#  e.g. ../e3_my_entities.json
+#  e.g. ../e3_my_entities.json. Files that are used in exercises and files
+#  that are used in solutions are different from each other so be careful
+#  when working with them. You can of course change the paths as you wish,
+#  but it is recommended to use the already given path names.
 WRITE_ENTITIES_FILEPATH = Path("../e3_context_entities_solution_entities.json")
 
 # ## Main script
