### 2024-04-16 01b6a90
chore: fix small typos and pep-8 errors in exercise 5 and 6
