+++ tutorials/ngsi_v2/e6_timeseries_data/e6_timeseries_data.py
@@ -150,6 +150,7 @@ if __name__ == '__main__':
     # Implement a callback function that gets triggered when the
     # command is sent to the device. The incoming command should update the
     # heater attribute of the simulation model
+
     def on_command(client, obj, msg):
         """
         Callback for incoming commands
@@ -202,7 +203,6 @@ if __name__ == '__main__':
     mqttc.message_callback_add(sub=TOPIC_CONTROLLER,
                                callback=on_measurement)
 
-
     # ToDo: Create http subscriptions that get triggered by updates of your
     #  device attributes. Note that you can also post the same subscription
     #  by the context broker.
@@ -214,7 +214,7 @@ if __name__ == '__main__':
         notification=Notification(**{
             'http': {'url': 'http://quantumleap:8668/v2/notify'}
         }),
-        throttling = 0)
+        throttling=0)
     )
 
     cbc.post_subscription(subscription=Subscription(
@@ -225,7 +225,7 @@ if __name__ == '__main__':
         notification=Notification(**{
             'http': {'url': 'http://quantumleap:8668/v2/notify'}
         }),
-        throttling = 0)
+        throttling=0)
     )
 
     cbc.post_subscription(subscription=Subscription(
@@ -236,7 +236,7 @@ if __name__ == '__main__':
         notification=Notification(**{
             'http': {'url': 'http://quantumleap:8668/v2/notify'}
         }),
-        throttling = 0)
+        throttling=0)
     )
 
     # connect to the mqtt broker and subscribe to your topic
