+++ tutorials/ngsi_v2/e5_iot_thermal_zone_control/e5_iot_thermal_zone_control_solution.py
@@ -38,6 +38,7 @@ from typing import List
 from urllib.parse import urlparse
 from uuid import uuid4
 import paho.mqtt.client as mqtt
+from pydantic import parse_file_as
 import matplotlib.pyplot as plt
 
 # import from filip
@@ -122,19 +123,10 @@ if __name__ == '__main__':
     history_weather_station = []
     history_zone_temperature_sensor = []
     history_heater = []
-
-    with open(READ_GROUPS_FILEPATH, 'r') as groups_file, open(READ_DEVICES_FILEPATH, 'r') as devices_file:
-        groups_list = json.loads(groups_file.read())
-        devices_list = json.loads(devices_file.read())
     
     # Create clients and restore devices and groups from file
-    groups: List[ServiceGroup] = []
-    devices: List[Device] = []
-    for group_dict in groups_list:
-        groups.append(ServiceGroup(**group_dict))
-    for device_dict in devices_list:
-        devices.append(Device(**device_dict))
-    
+    groups = parse_file_as(List[ServiceGroup], READ_GROUPS_FILEPATH)
+    devices = parse_file_as(List[Device], READ_DEVICES_FILEPATH)
     cbc = ContextBrokerClient(url=CB_URL, fiware_header=fiware_header)
     iotac = IoTAClient(url=IOTA_URL, fiware_header=fiware_header)
     iotac.post_groups(service_groups=groups)
