### 2024-04-17 98cfced
Added update_entity methods for all attributes and relationships
