### 2024-04-17 eac5243
Added test for update_entity_properties and update_entity_relationships
