### 2024-04-17 b3da0b3
chore: finish the cb_entity and test get_properties
