### 2024-04-17 8e2e9d9
chore: finish the cb_entity and test get_properties
