### 2024-04-18 6cb8ac9
chore: update to paho v2 and add geojson requirement #270
