+++ examples/ngsi_v2/e08_ngsi_v2_iota_paho_mqtt.py
@@ -207,6 +207,7 @@ if __name__ == '__main__':
     mqtt_client = mqtt.Client(client_id="filip-iot-example",
                               userdata=None,
                               protocol=mqtt.MQTTv5,
+                              callback_api_version=mqtt.CallbackAPIVersion.VERSION1,
                               transport="tcp")
     # add our callbacks to the client
     mqtt_client.on_connect = on_connect
