### 2024-04-18 7c531d9
chore: update to paho v2 and add geojson requirement #270
