+++ requirements.txt
@@ -16,7 +16,7 @@ python-Levenshtein~=0.23.0
 python-dateutil~=2.8.2
 wget~=3.2
 stringcase~=1.2.0
-paho-mqtt~=1.6.1
+paho-mqtt>=2.0.0
 datamodel_code_generator[http]~=0.25.0
 # tutorials
 matplotlib~=3.5.3; python_version < '3.9'
