### 2024-04-18 ab8cfb4
chore: add get_context method + unittest
