### 2024-04-23 31ad393
fix: add geojson_pydantic in setup.py
