+++ setup.py
@@ -18,6 +18,7 @@ INSTALL_REQUIRES = ['aenum~=3.1.15',
                     'regex~=2023.10.3',
                     'requests~=2.31.0',
                     'rapidfuzz~=3.4.0',
+                    'geojson_pydantic~=1.0.2',
                     'wget~=3.2']
 
 SETUP_REQUIRES = INSTALL_REQUIRES.copy()
