+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.4.1
 - fix: Session added as optional parameter to enable tls communication with clients ([#249](https://github.com/RWTH-EBC/FiLiP/pull/249))
+- fix: add missing package ``geojson_pydantic`` in setup.py
 
 #### v0.4.0
 - add tutorial for protected endpoint with bearer authentication ([#208](https://github.com/RWTH-EBC/FiLiP/issues/208))
