### 2024-04-23 7034f1a
chore: add clean up at the start
