+++ examples/ngsi_v2/e12_ngsi_v2_use_case_models.py
@@ -106,6 +106,8 @@ if __name__ == "__main__":
     # 0. Initial client
     cb_client = ContextBrokerClient(url=CB_URL,
                                     fiware_header=fiware_header)
+    # clear cb
+    clear_context_broker(cb_client=cb_client)
 
     # 1. Crate data
     weather_station = WeatherStationFIWARE(
@@ -151,4 +153,4 @@ if __name__ == "__main__":
           f"{weather_station_2_fiware.model_dump_json(indent=2, include={'id', 'type'})}\n")
 
     # clear cb
-    clear_context_broker(url=CB_URL, fiware_header=fiware_header)
+    clear_context_broker(cb_client=cb_client)
