### 2024-04-23 8e27c6e
chore: reform assertRaises
