### 2024-04-23 0cfc7d4
Added more tests for (JEXL) expression language and fixed attribute types in Tutorial e8
