+++ tutorials/ngsi_v2/e8_multientity_and_expression_language/e8_multientity_and_expression_language_solution.py
@@ -93,8 +93,8 @@ if __name__ == '__main__':
 
     # TODO: Setting expression language to JEXL at Device level with other attributes.
     #  The attribute 'value' (Number) is itself multiplied by 5. The attribute
-    #  'consumption' (String) is the trimmed version of the attribute 'spaces' (String).
-    #  The attribute 'iso_time' (String) is the current 'timestamp' (Number) transformed into the ISO format.
+    #  'consumption' (Text) is the trimmed version of the attribute 'spaces' (Text).
+    #  The attribute 'iso_time' (Text) is the current 'timestamp' (Number) transformed into the ISO format.
     device2 = Device(device_id="waste_container_002",
                      entity_name="urn:ngsi-ld:WasteContainer:002",
                      entity_type="WasteContainer",
@@ -103,11 +103,11 @@ if __name__ == '__main__':
                      expressionLanguage=ExpressionLanguage.JEXL,
                      attributes=[DeviceAttribute(name="value", type="Number",
                                                  expression="5 * value"),
-                                 DeviceAttribute(name="spaces", type="String"),
-                                 DeviceAttribute(name="consumption", type="String",
+                                 DeviceAttribute(name="spaces", type="Text"),
+                                 DeviceAttribute(name="consumption", type="Text",
                                                  expression="spaces|trim"),
                                  DeviceAttribute(name="timestamp", type="Number"),
-                                 DeviceAttribute(name="iso_time", type="String",
+                                 DeviceAttribute(name="iso_time", type="Text",
                                                  expression="timestamp|toisodate"),
                                  ]
                      )
