### 2024-04-23 f854572
Optimized tests for jexl and legacy expression language
