### 2024-04-23 e55dfe0
Added pyjexl package to setup.py
