+++ setup.py
@@ -23,7 +23,7 @@ INSTALL_REQUIRES = ['aenum~=3.1.15',
 
 SETUP_REQUIRES = INSTALL_REQUIRES.copy()
 
-VERSION = '0.4.0'
+VERSION = '0.4.1'
 
 setuptools.setup(
     name='filip',
