### 2024-04-24 20c7062
chore: add test cases for subscription model
