### 2024-04-24 9a2c3b4
chore: add hints to transformation functions
