### 2024-04-24 a6e1e91
chore: tiny adjustment of the tutorials
