+++ tutorials/ngsi_v2/e8_multientity_and_expression_language/e8_multientity_and_expression_language.py
@@ -1,8 +1,8 @@
 """
 # # Exercise 8: MultiEntity and Expression Language
 
-# The MultiEntity plugin Allows the devices provisioned in the IoTAgent to map their
-attributes to more than one entity, # declaring the target entity through the
+# The MultiEntity plugin allows the devices provisioned in the IoTAgent to map their
+attributes to more than one entity, declaring the target entity through the
 Configuration or Device provisioning APIs.
 
 # The IoTAgent Library provides an expression language for measurement transformation,
@@ -27,6 +27,7 @@ import datetime
 
 from filip.clients.ngsi_v2 import IoTAClient, ContextBrokerClient
 from filip.models.base import FiwareHeader
+from filip.models.ngsi_v2.context import ContextEntity, NamedContextAttribute
 from filip.models.ngsi_v2.iot import (Device, ServiceGroup, TransportProtocol,
                                       PayloadProtocol, DeviceAttribute,
                                       ExpressionLanguage)
@@ -77,8 +78,9 @@ if __name__ == '__main__':
 
     # TODO: Create a device with two attributes 'location' and 'fillingLevel' that use
     #  expressions. These attributes are based on the attributes 'longitude',
-    #  'latitude' and 'level'. 'location' is an array with 'longitude' and 'latitude'.
-    #  'fillingLevel' is 'level' divided by 100
+    #  'latitude' and 'level', while:
+    #  1. 'location' is an array with 'longitude' and 'latitude'.
+    #  2. 'fillingLevel' is 'level' divided by 100
     device1 = Device(device_id="waste_container_001",
                      entity_name="urn:ngsi-ld:WasteContainer:001",
                      entity_type="WasteContainer",
@@ -88,10 +90,10 @@ if __name__ == '__main__':
                      )
     iota_client.post_device(device=device1)
 
-    # TODO: Setting expression language to JEXL at Device level with other attributes.
-    #  The attribute 'value' (Number) is itself multiplied by 5. The attribute
-    #  'consumption' (Text) is the trimmed version of the attribute 'spaces' (Text).
-    #  The attribute 'iso_time' (Text) is the current 'timestamp' (Number) transformed into the ISO format.
+    # TODO: Setting expression language to JEXL at Device level with five attributes, while
+    #  1. The attribute 'value' (Number) is itself multiplied by 5. The attribute
+    #  2. 'consumption' (Text) is the trimmed version of the attribute 'spaces' (Text).
+    #  3. The attribute 'iso_time' (Text) is the current 'timestamp' (Number) transformed into the ISO format.
     device2 = Device(device_id="waste_container_002",
                      entity_name="urn:ngsi-ld:WasteContainer:002",
                      entity_type="WasteContainer",
@@ -114,10 +116,26 @@ if __name__ == '__main__':
 
     client.disconnect()
 
-    # TODO: Create a weather station device with multi entity attributes (Number). 'v'
-    #  is multiplied by 100 and is a standard attribute. 'v1' and 'v2' are multiplied
-    #  by 100 and are multi entity attributes of type SubWeatherStation. The name of
-    #  each attribute is 'vol'.
+    # Printing context entities of OCB
+    for context_entity in cb_client.get_entity_list(entity_types=["WasteContainer"]):
+        print(context_entity.model_dump_json(indent=4))
+
+    # Creating two SubWeatherStation entities
+    entity1 = ContextEntity(id="urn:ngsi-ld:SubWeatherStation:001",
+                            type="SubWeatherStation")
+    entity1.add_attributes(attrs=[NamedContextAttribute(name="vol", type="Number")])
+    cb_client.post_entity(entity1)
+
+    entity2 = ContextEntity(id="urn:ngsi-ld:SubWeatherStation:002",
[…]
