+++ tutorials/ngsi_v2/e8_multientity_and_expression_language/README.md
@@ -5,9 +5,9 @@ declaring the target entity through the Configuration or Device provisioning API
 
 The IoTAgent Library provides an expression language for measurement transformation, that can be used to adapt the
 information coming from the South Bound APIs to the information reported to the Context Broker. This is really useful
-when you need to adapt measure.
+when you need to adapt measured data.
 
-There are available two different expression languages jexl and legacy. The recommended language to use is jexl,
+There are two different expression languages available, `jexl` and `legacy`. The recommended language is `jexl`,
 which is newer and most powerful.
 
 We want to apply these features to `e8_multientity_and_expression_language.py`.
