### 2024-04-25 fbc7187
Added registrations argument to clear context broker, refined test
