+++ tests/clients/test_ngsi_v2_iota.py
@@ -403,7 +403,7 @@ class TestAgent(unittest.TestCase):
         with self.assertRaises(KeyError):
             self.client.get_group(resource="/iot/json", apikey="not_exist")
 
-        self.tearDown()
+        #self.tearDown()
 
     def test_update_service_group(self):
         """
@@ -442,22 +442,42 @@ class TestAgent(unittest.TestCase):
         Returns:
             None
         """
+        cb_client = ContextBrokerClient(url=settings.CB_URL,
+                                        fiware_header=self.fiware_header)
         device = Device(**self.device)
         device.add_command(DeviceCommand(name="dummy_cmd"))
         self.client.post_device(device=device)
-        clear_context_broker(settings.CB_URL,self.fiware_header)
-        self.assertRaises(requests.HTTPError,clear_iot_agent,
-                          settings.IOTA_URL,self.fiware_header)
+        clear_context_broker(settings.CB_URL,
+                             self.fiware_header)
+        self.assertEqual(len(cb_client.get_registration_list()),1)
 
+        clear_iot_agent(settings.IOTA_JSON_URL,self.fiware_header)
+        self.assertCountEqual(cb_client.get_registration_list(),[])
+
+        self.client.post_device(device=device)
+        clear_context_broker(settings.CB_URL,
+                             self.fiware_header,
+                             clear_registrations=True)
+        with self.assertRaises(requests.HTTPError):
+            clear_iot_agent(settings.IOTA_URL, self.fiware_header)
+    
     def tearDown(self) -> None:
         """
         Cleanup test server
 
         """
-        self.client.close()
-        try:
+        cb_client = ContextBrokerClient(url=settings.CB_URL,
+                                        fiware_header=self.fiware_header)
+        devs_with_cmds = [dev.commands for dev in self.client.get_device_list() 
+                          if dev.commands !=[]]
+        if devs_with_cmds != [] and cb_client.get_registration_list() == []:
+            print("Dangling device with command and no registration found")
+            clear_context_broker(url=settings.CB_URL,
+                                 fiware_header=self.fiware_header)
+        else:
+            self.client.close()
             clear_all(fiware_header=self.fiware_header,
-                    cb_url=settings.CB_URL,
-                    iota_url=settings.IOTA_JSON_URL)
-        except requests.HTTPError:
-            print("Already cleared")
+                cb_url=settings.CB_URL,
+                iota_url=settings.IOTA_JSON_URL)
+            
+        
