### 2024-04-26 62c870d
run all entity tests
Signed-off-by: iripiri <iris.m.koester@gmail.com>
