+++ filip/models/ngsi_ld/context.py
@@ -581,7 +581,7 @@ class ContextLDEntity(ContextLDEntityKeyValues):
             None
         """
         if isinstance(attrs, list):
-            attrs = {attr.name: ContextProperty(**attr.dict(exclude={'name'}))
+            attrs = {attr.name: ContextProperty(**attr.model_dump(exclude={'name'}))
                      for attr in attrs}
         for key, attr in attrs.items():
             self.__setattr__(name=key, value=attr)
