### 2024-04-26 2fd8dfc
run all entity tests
Signed-off-by: iripiri <iris.m.koester@gmail.com>
