### 2024-04-30 bc26f9e
fix: change the order of the clearing functions
