+++ tutorials/ngsi_v2/e4_iot_thermal_zone_sensors/e4_iot_thermal_zone_sensors_solution.py
@@ -83,8 +83,8 @@ if __name__ == '__main__':
     fiware_header = FiwareHeader(service=SERVICE,
                                  service_path=SERVICE_PATH)
     # clear the state of your service and scope
-    clear_context_broker(url=CB_URL, fiware_header=fiware_header)
     clear_iot_agent(url=IOTA_URL, fiware_header=fiware_header)
+    clear_context_broker(url=CB_URL, fiware_header=fiware_header)
 
     # instantiate simulation model
     sim_model = SimulationModel(t_start=T_SIM_START,
@@ -282,5 +282,5 @@ if __name__ == '__main__':
         groups = [item.model_dump() for item in iotac.get_group_list()]
         json.dump(groups, f, ensure_ascii=False, indent=2)
 
-    clear_context_broker(url=CB_URL, fiware_header=fiware_header)
     clear_iot_agent(url=IOTA_URL, fiware_header=fiware_header)
+    clear_context_broker(url=CB_URL, fiware_header=fiware_header)
