### 2024-04-30 4e4b00e
fix: remove the test of deleting registration before device
