### 2024-04-30 8317608
fix: revert changes to exercise 6
