+++ tutorials/ngsi_v2/e6_timeseries_data/e6_timeseries_data_solution.py
@@ -32,7 +32,8 @@ from filip.clients.ngsi_v2 import \
 from filip.clients.mqtt import IoTAMQTTClient
 from filip.models.base import FiwareHeader
 from filip.models.ngsi_v2.context import NamedCommand
-from filip.models.ngsi_v2.subscriptions import Subscription, Message
+from filip.models.ngsi_v2.subscriptions import Subscription, Message, Notification, \
+    Subject
 from filip.models.ngsi_v2.iot import \
     Device, \
     PayloadProtocol, \
@@ -202,29 +203,42 @@ if __name__ == '__main__':
     mqttc.message_callback_add(sub=TOPIC_CONTROLLER,
                                callback=on_measurement)
 
-    # ToDo: Create a quantumleap client.
-    qlc = QuantumLeapClient(url=QL_URL, fiware_header=fiware_header)
-
     # ToDo: Create http subscriptions that get triggered by updates of your
     #  device attributes. Note that you can only post the subscription
     #  to the context broker.
-    qlc.post_subscription(entity_id=weather_station.entity_name,
-                          entity_type=weather_station.entity_type,
-                          cb_url="http://orion:1026",
-                          ql_url="http://quantumleap:8668",
-                          throttling=0)
 
-    qlc.post_subscription(entity_id=zone_temperature_sensor.entity_name,
-                          entity_type=zone_temperature_sensor.entity_type,
-                          cb_url="http://orion:1026",
-                          ql_url="http://quantumleap:8668",
-                          throttling=0)
+    cbc.post_subscription(subscription=Subscription(
+        subject=Subject(**{
+            'entities': [{'id': weather_station.entity_name,
+                          'type': weather_station.entity_type}]
+        }),
+        notification=Notification(**{
+            'http': {'url': 'http://quantumleap:8668/v2/notify'}
+        }),
+        throttling=0)
+    )
 
-    qlc.post_subscription(entity_id=heater.entity_name,
-                          entity_type=heater.entity_type,
-                          cb_url="http://orion:1026",
-                          ql_url="http://quantumleap:8668",
-                          throttling=0)
+    cbc.post_subscription(subscription=Subscription(
+        subject=Subject(**{
+            'entities': [{'id': zone_temperature_sensor.entity_name,
+                          'type': zone_temperature_sensor.entity_type}]
+        }),
+        notification=Notification(**{
+            'http': {'url': 'http://quantumleap:8668/v2/notify'}
+        }),
+        throttling=0)
+    )
+
+    cbc.post_subscription(subscription=Subscription(
+        subject=Subject(**{
+            'entities': [{'id': heater.entity_name,
+                          'type': heater.entity_type}]
+        }),
+        notification=Notification(**{
+            'http': {'url': 'http://quantumleap:8668/v2/notify'}
+        }),
+        throttling=0)
+    )
 
     # connect to the mqtt broker and subscribe to your topic
     mqtt_url = urlparse(MQTT_BROKER_URL_EXPOSED)
@@ -278,6 +292,9 @@ if __name__ == '__main__':
 
     # wait until all data is available
     time.sleep(10)
+
+    # ToDo: Create a quantumleap client.
+    qlc = QuantumLeapClient(url=QL_URL, fiware_header=fiware_header)
 
     # ToDo: Retrieve the historic data from QuantumLeap, convert them to a
     #  pandas dataframe and plot them.
