### 2024-04-30 9cebb59
fix: fix deprecated methods in e05
