### 2024-04-30 cf7d371
format: reformat the sentences and fix typos in e05
