### 2024-04-30 441f343
fix: reverse kilometer to kilometre for pydantic
