### 2024-04-30 8a9a747
chore: remove some examples in e6
