### 2024-04-30 ab4c22b
feat: add test for device attributes
