### 2024-05-01 60ff6b9
fix: fix deprecated methods in e08
