### 2024-05-07 59fd7ae
feat: add validation of duplicated attributes
