### 2024-05-07 3c3ffd2
chore: test device attributes during creation
