### 2024-05-07 291c0b8
feat: remove milliseconds from logger
