### 2024-05-07 31cafa0
feat: change print statements for better readability in e10
