### 2024-05-07 fbe7576
fix: remove redundant subscription_id
