### 2024-05-07 0cfa4cd
chore: delete all matched devices if no entity type is given
