### 2024-05-08 23ecb6b
fix: fix typos and pep 8 errors in e07
