### 2024-05-08 bee4278
feat: remove milliseconds from logger and use only printf statements in e07
