### 2024-05-08 5ff889e
fix: fix f string in cb.py
