### 2024-05-08 ea33441
changed result of attr endpoint methods and added test
