+++ filip/clients/ngsi_v2/quantumleap.py
@@ -2,10 +2,11 @@
 TimeSeries Module for QuantumLeap API Client
 """
 import logging
+import numpy as np
 import time
 from math import inf
 from collections import deque
-from itertools import count
+from itertools import count,chain
 from typing import Dict, List, Union, Deque, Optional
 from urllib.parse import urljoin
 import requests
@@ -21,10 +22,7 @@ from filip.models.ngsi_v2.timeseries import \
     AggrScope, \
     AttributeValues, \
     TimeSeries, \
-    TimeSeriesHeader, \
-    TimeSeriesAttrHeader, \
-    TimeSeriesEntities, \
-    EntityValues
+    TimeSeriesHeader
 from filip.utils.validators import validate_http_url
 
 
@@ -1064,45 +1062,28 @@ class QuantumLeapClient(BaseHttpClient):
         """
         url = urljoin(self.base_url, f'v2/types/{entity_type}/attrs/'
                                      f'{attr_name}/value')
-        res_q = self.__query_builder(url=url,
-                                     entity_id=entity_id,
-                                     id_pattern=id_pattern,
-                                     options=options,
-                                     entity_type=entity_type,
-                                     aggr_method=aggr_method,
-                                     aggr_period=aggr_period,
-                                     from_date=from_date,
-                                     to_date=to_date,
-                                     last_n=last_n,
-                                     limit=limit,
-                                     offset=offset,
-                                     georel=georel,
-                                     geometry=geometry,
-                                     coords=coords,
-                                     aggr_scope=aggr_scope)
-
-        # merge chunks of response
-        res = [TimeSeries(index=item.get('index'),
-                          entityType=entity_type,
-                          entityId=item.get('entityId'),
-                          attributes=[
-                              AttributeValues(attrName=attr_name,
-                                              values=item.get('values'))])
-               for item in res_q.popleft().get('values')]
-
-        for chunk in res_q:
-            chunk = [TimeSeries(index=item.get('index'),
-                                entityType=entity_type,
-                                entityId=item.get('entityId'),
-                                attributes=[
-                                    AttributeValues(attrName=attr_name,
-                                                    values=item.get('values'))])
-                     for item in chunk.get('values')]
-
-            for new, old in zip(chunk, res):
-                old.extend(new)
-
-        return res
+        res = self.__query_builder(url=url,
+                                   entity_id=entity_id,
+                                   options=options,
+                                   entity_type=entity_type,
+                                   aggr_method=aggr_method,
+                                   aggr_period=aggr_period,
+                                   from_date=from_date,
+                                   to_date=to_date,
+                                   last_n=last_n,
+                                   limit=limit,
+                                   offset=offset,
+                                   georel=georel,
+                                   geometry=geometry,
+                                   coords=coords,
+                                   aggr_scope=aggr_scope)
+        return [TimeSeries(index=item.get('index'),
+                           entityType=entity_type,
+                           entityId=item.get('entityId'),
+                           attributes=[
+                               AttributeValues(attrName=attr_name,
+                                               values=item.get('values'))])
+                for item in res.get('values')]
 
     # v2/attrs
     def get_entity_by_attrs(self, *,
@@ -1111,7 +1092,7 @@ class QuantumLeapClient(BaseHttpClient):
                             to_date: str = None,
                             limit: int = None,
                             offset: int = None
-                            ) -> List[TimeSeriesAttrHeader]:
+                            ) -> List[TimeSeries]:
         """
         Get list of timeseries data grouped by each existing attribute name.
         The timeseries data include all entities corresponding to each
@@ -1133,29 +1114,30 @@ class QuantumLeapClient(BaseHttpClient):
[…]
