### 2024-05-08 912a48b
fix: fix logger output in e01
