### 2024-05-15 e753eab
fixes after datamodel changes
Signed-off-by: iripiri <iris.m.koester@gmail.com>
