### 2024-05-15 3b9a499
Added mqqt endpoint for subscriptions.
