### 2024-05-15 b41b7e4
chore: update the get_context method in order to respect entitites without context
