+++ tests/clients/test_mqtt_client.py
@@ -61,16 +61,16 @@ class TestMQTTClient(unittest.TestCase):
 
         self.mqttc = IoTAMQTTClient()
 
-        def on_connect(mqttc, obj, flags, rc):
+        def on_connect(mqttc, obj, flags, rc,properties):
             mqttc.logger.info("rc: " + str(rc))
 
         def on_connect_fail(mqttc, obj):
             mqttc.logger.info("Connect failed")
 
-        def on_publish(mqttc, obj, mid):
+        def on_publish(mqttc, obj, mid,rc,properties):
             mqttc.logger.info("mid: " + str(mid))
 
-        def on_subscribe(mqttc, obj, mid, granted_qos):
+        def on_subscribe(mqttc, obj, mid, granted_qos,properties):
             mqttc.logger.info("Subscribed: " + str(mid)
                               + " " + str(granted_qos))
 
