+++ examples/ngsi_v2/e08_ngsi_v2_iota_paho_mqtt.py
@@ -201,7 +201,7 @@ if __name__ == '__main__':
                              f"/{device.device_id}/cmdexe",
                        payload=json.dumps(res))
 
-    def on_disconnect(client, userdata, reasonCode, properties):
+    def on_disconnect(client, userdata, flags, reasonCode, properties):
         logger.info("MQTT client disconnected" + str(reasonCode))
 
     mqtt_client = mqtt.Client(client_id="filip-iot-example",
