### 2024-05-15 0fdaac5
Switched CallbackAPI VERSION1 settings to VERSION2
