### 2024-05-16 53e578b
feature: reformat logger date and remove milliseconds
