+++ examples/ngsi_v2/e10_ngsi_v2_quantumleap.py
@@ -31,8 +31,7 @@ SERVICE_PATH = '/example'
 logging.basicConfig(
     level='INFO',
     format='%(asctime)s %(name)s %(levelname)s: %(message)s',
-    datefmt='%Y-%m-%d %H:%M:%S')
-
+    datefmt='%d-%m-%Y %H:%M:%S')
 logger = logging.getLogger(__name__)
 
 if __name__ == "__main__":
