+++ examples/ngsi_v2/e03_ngsi_v2_context_subscriptions_http.py
@@ -37,7 +37,8 @@ SERVER_URL = "http://<my_url>/notify/"
 # Setting up logging
 logging.basicConfig(
     level='INFO',
-    format='%(asctime)s %(name)s %(levelname)s: %(message)s')
+    format='%(asctime)s %(name)s %(levelname)s: %(message)s',
+    datefmt='%d-%m-%Y %H:%M:%S')
 logger = logging.getLogger(__name__)
 
 
