+++ examples/ngsi_v2/e09_ngsi_v2_iota_filip_mqtt.py
@@ -47,7 +47,8 @@ SERVICE_GROUP_APIKEY = 'filip-example-service-group'
 # Setting up logging
 logging.basicConfig(
     level='INFO',
-    format='%(asctime)s %(name)s %(levelname)s: %(message)s')
+    format='%(asctime)s %(name)s %(levelname)s: %(message)s',
+    datefmt='%d-%m-%Y %H:%M:%S')
 
 logger = logging.getLogger(__name__)
 
