### 2024-05-21 a93d735
format: reformat mqtt notifications in e04
