+++ tests/clients/test_ngsi_v2_timeseries.py
@@ -399,16 +399,16 @@ class TestTimeSeries(unittest.TestCase):
         with QuantumLeapClient(
                 url=settings.QL_URL,
                 fiware_header=FiwareHeader(service='filip',
-                                 service_path="/static")) \
+                                           service_path="/static")) \
                 as client:
-            attr_names = ['temperature','humidity','co2']
+            attr_names = ['temperature', 'humidity', 'co2']
             for attr_name in attr_names:
                 entities_by_attr_name = client.get_entity_by_attr_name(
                     attr_name=attr_name) 
-                #we expect as many timeseries as there are unique ids
+                # we expect as many timeseries as there are unique ids
                 self.assertEqual(len(entities_by_attr_name), 2) 
 
-                #we expect the sizes of the index and attribute values to be the same
+                # we expect the sizes of the index and attribute values to be the same
                 for timeseries in entities_by_attr_name:
                     for attribute in timeseries.attributes:
                         self.assertEqual(len(attribute.values),len(timeseries.index))
