+++ examples/ngsi_v2/e05_ngsi_v2_context_registrations.py
@@ -27,7 +27,7 @@ CB_URL = "http://localhost:1026"
 # You can also change the used Fiware service
 # FIWARE-Service
 SERVICE = 'filip'
-# FIWARE-Servicepath
+# FIWARE-Service path
 SERVICE_PATH = '/example'
 
 # Setting up logging
@@ -40,7 +40,7 @@ logger = logging.getLogger(__name__)
 if __name__ == '__main__':
     # # 1 Creating models
     #
-    # Create a building with a weather station as context provider.
+    # Create a building with a weather station as a context provider.
     # We start by creating metadata. The other way round is also possible, but
     # you have to use the api of the Context Entity Model.
 
@@ -48,22 +48,20 @@ if __name__ == '__main__':
     temperature_metadata = NamedMetadata(name="unit",
                                          type="Unit",
                                          value=Unit(name="degree Celsius").model_dump())
-    # create the temperature attribute of the weather station
+    # create the 'temperature' attribute of the weather station
     temperature = ContextAttribute(type="Number",
                                    value=20.5,
                                    metadata=temperature_metadata)
-    # create the complete model of the building with the weather station
-    weather_station_building = ContextEntity(id="urn:ngsi-ld:WeatherStation:001",
-                                             type="WeatherStation",
-                                             temperature=temperature)
+    # create the complete model of the weather station
+    weather_station = ContextEntity(id="urn:ngsi-ld:WeatherStation:001",
+                                    type="WeatherStation",
+                                    temperature=temperature)
 
     # print the complete weather station object
-    print("+" * 80)
-    print("Building with weather station with one property from a sensor")
-    print("+" * 80)
-    print(weather_station_building.model_dump_json(indent=2))
+    print(f"{'*' * 80}\nWeather station with one attribute\n{'*' * 80}")
+    print(weather_station.model_dump_json(indent=2))
 
-    # create additional properties of the weather station
+    # create an additional attribute 'wind_speed' of the weather station
     wind_speed_metadata = NamedMetadata(name="unit",
                                         type="Unit",
                                         value=Unit(name="kilometre per "
@@ -72,21 +70,17 @@ if __name__ == '__main__':
     wind_speed = ContextAttribute(type="Number",
                                   value=60,
                                   metadata=wind_speed_metadata)
-    weather_station_building.add_attributes(attrs={"wind_speed": wind_speed})
+    weather_station.add_attributes(attrs={"wind_speed": wind_speed})
 
     # print the complete model
-    print("+" * 80)
-    print("Building with weather station with two properties from a sensor")
-    print("+" * 80)
-    print(weather_station_building.model_dump_json(indent=2))
+    print(f"{'*' * 80}\nWeather station with two attributes\n{'*' * 80}")
+    print(weather_station.model_dump_json(indent=2))
 
-    building_1 = ContextEntity(id="urn:ngsi-ld:building:001",
-                               type="Building")
+    building = ContextEntity(id="urn:ngsi-ld:building:001",
+                             type="Building")
 
-    print("+" * 80)
-    print("Building without own properties")
-    print("+" * 80)
-    print(building_1.model_dump_json(indent=2))
+    print(f"{'*' * 80}\nBuilding without own attributes\n{'*' * 80}")
+    print(building.model_dump_json(indent=2))
 
     # # 2 Creating registration
     #
@@ -107,38 +101,35 @@ if __name__ == '__main__':
             },
             "provider": provider.model_dump()
         })
-    print("+" * 80)
-    print("Registration that makes the first building a context "
-          "provider for the second building")
-    print("+" * 80)
[…]
