### 2024-05-24 07ff87e
fix: Fixing validation of cbHost attribute. Also updated the tutorials with the new functionianlity of pydantic (except tutorial e6).
