### 2024-05-24 b0f14f0
add: Providing Docker compose file that works for the tutorials.
