+++ tutorials/ngsi_v2/fiware_env/docker-compose.yml
@@ -1,0 +1,96 @@
+
+volumes:
+    mongodb: ~
+    crate-db: ~
+
+services:
+    iot-agent:
+        image: fiware/iotagent-json:1.26.0
+        hostname: iot-agent
+        container_name: fiware-iot-agent
+        depends_on:
+            - mongodb
+            - mosquitto
+        expose:
+            - "4041"
+        ports:
+            - "4041:4041"
+        environment:
+            - "IOTA_CB_HOST=orion"
+            - "IOTA_CB_PORT=1026"
+            - "IOTA_NORTH_PORT=4041"
+            - "IOTA_REGISTRY_TYPE=mongodb"
+            - "IOTA_MONGO_HOST=mongodb"
+            - "IOTA_MONGO_PORT=27017"
+            - "IOTA_MONGO_DB=iotagent-json"
+            - "IOTA_PROVIDER_URL=http://iot-agent:4041"
+            - "IOTA_MQTT_HOST=mosquitto"
+            - "IOTA_MQTT_PORT=1883"
+            - "IOTA_DEFAULT_RESOURCE="
+            - "IOTA_DEFAULT_TRANSPORT=MQTT"
+            - "IOTA_LOG_LEVEL=DEBUG"
+            - "IOTA_TIMESTAMP=true"
+            - "IOTA_CB_NGSI_VERSION=v2"
+            - "IOTA_AUTOCAST=true"
+
+    mongodb:
+        image: mongo:4.4
+        hostname: mongodb
+        container_name: db-mongo
+        ports:
+            - "27017:27017"
+        volumes:
+            - mongodb:/data
+
+    orion:
+        image: fiware/orion:3.7.0
+        hostname: orion
+        container_name: fiware-orion
+        depends_on:
+            - mongodb
+        expose:
+            - "1026"
+        ports:
+            - "1026:1026"
+        command: 
+          -dbhost mongodb
+          -logLevel DEBUG
+          -logForHumans
+
+    mosquitto:
+        image: eclipse-mosquitto:2.0.18
+        hostname: mosquitto
+        container_name: mosquitto
+        expose:
+          - "1883"
+          - "9001"
+        ports:
+          - "1883:1883"
+          - "9001:9001"
+        volumes:
+          - ./mosquitto.conf:/mosquitto/config/mosquitto.conf
+
+    crate-db:
+        image: crate:5.4.4
+        container_name: cratedb
+        hostname: crate-db
+        ports:
+            - '4200:4200'
+            - '4300:4300'
+        command:
+            crate -Cauth.host_based.enabled=false  -Ccluster.name=democluster -Chttp.cors.enabled=true
+            -Chttp.cors.allow-origin="*"
+        environment:
+            - CRATE_HEAP_SIZE=2g
+
+    quantumleap:
+        image: orchestracities/quantumleap:0.8.3
+        container_name: quantumleap
+        hostname: quantumleap
+        ports:
+            - '8668:8668'
+        depends_on:
+            - crate-db
+        environment:
+            - CRATE_HOST=crate-db
+            - LOGLEVEL=DEBUG
