### 2024-06-05 d46f296
fix: fix difference between logger and actual data in fiware
