### 2024-06-05 54b4cf0
format: fix typos in e08
