+++ examples/ngsi_v2/e08_ngsi_v2_iota_paho_mqtt.py
@@ -24,7 +24,7 @@ from filip.clients.ngsi_v2 import HttpClient, HttpClientConfig
 # ## Parameters
 #
 # To run this example you need a working Fiware v2 setup with a
-# context-broker, an IoT-Agent and mqtt-broker. You can here set the
+# context-broker, an IoT-Agent and mqtt-broker. Here you can set the
 # addresses:
 #
 # Host address of Context Broker
@@ -34,10 +34,10 @@ IOTA_URL = "http://localhost:4041"
 # Host address of the MQTT-Broker
 MQTT_BROKER_URL = "mqtt://localhost:1883"
 
-# You can here also change the used Fiware service
+# Here you can also change the used Fiware service
 # FIWARE-Service
 SERVICE = 'filip'
-# FIWARE-Servicepath
+# FIWARE-Service path
 SERVICE_PATH = '/example'
 
 # You may also change the ApiKey Information
@@ -58,7 +58,7 @@ if __name__ == '__main__':
 
     # # 1 FiwareHeader
     # Since we want to use the multi-tenancy concept of fiware we always start
-    # with create a fiware header
+    # with creating a fiware header
     fiware_header = FiwareHeader(service=SERVICE,
                                  service_path=SERVICE_PATH)
 
@@ -69,9 +69,9 @@ if __name__ == '__main__':
     # First we create our device configuration using the models provided for
     # filip.models.ngsi_v2.iot
 
-    # creating an attribute for incoming measurements from e.g. a sensor we do
-    # add the metadata for units here using the unit models. You will later
-    # notice that the library automatically will augment the provided
+    # Creating an attribute for incoming measurements from e.g. a sensor. We
+    # add the metadata for units using the unit models. You will later
+    # notice that the library automatically augments the provided
     # information about units.
     device_attr1 = DeviceAttribute(name='temperature',
                                    object_id='t',
@@ -80,7 +80,7 @@ if __name__ == '__main__':
                                                  {"type": "Unit",
                                                   "value": {
                                                       "name": "degree Celsius"
-                                                  }
+                                                 }
                                                   }
                                              })
 
@@ -93,12 +93,12 @@ if __name__ == '__main__':
     device_command = DeviceCommand(name='heater')
 
     # NOTE: You need to know that if you define an apikey for a single device it
-    # will be only used for outgoing traffic. This is does not become very clear
+    # will be only used for outgoing traffic. This is not clearly defined
     # in the official documentation.
     # https://fiware-iotagent-json.readthedocs.io/en/latest/usermanual/index.html
-    device = Device(device_id='MyDevice',
-                    entity_name='MyDevice',
-                    entity_type='Thing2',
+    device = Device(device_id='urn:ngsi-ld:device:001',
+                    entity_name='urn:ngsi-ld:device:001',
+                    entity_type='Thing',
                     protocol='IoTA-JSON',
                     transport='MQTT',
                     apikey=DEVICE_APIKEY,
@@ -106,7 +106,7 @@ if __name__ == '__main__':
                     static_attributes=[static_device_attr],
                     commands=[device_command])
 
-    # You can also add additional attributes via the Device API
+    # you can also add additional attributes via the Device API
     device_attr2 = DeviceAttribute(name='humidity',
                                    object_id='h',
                                    type="Number",
@@ -116,67 +116,68 @@ if __name__ == '__main__':
 
     device.add_attribute(attribute=device_attr2)
 
-    # This will print our configuration that we will send
+    # this will print our configuration that we will send
     logging.info("This is our device configuration: \n" + device.model_dump_json(indent=2))
 
-    # ## 2.2 Device Submission
+    # ## 2.2 Device Provision
     #
[…]
