### 2024-06-11 205c1d4
Adding default resource /iot/json to docker compose file of tutorials
