### 2024-06-12 f2dc6eb
format: reformat mqtt callback functions in e04 and e08
