### 2024-06-12 ddb08b1
format: reformat mqtt callbacks and print statements in e09
