### 2024-06-12 c690a72
format: reformat on_message callback function in e04
