### 2024-06-12 #6538 @djs0109
> @djs0109 We will proceed with reviewing your submission, once we find an editor. However, could you work on revising the paper to make the research applications of your software more clear? This was the main feedback from our scope review.

Sure! We are working on it now.
