### 2024-06-12 b628a0a
chore: remove unused package in cb_test for v2
