### 2024-06-12 13fcf93
feat: add LD_CB_URL as new environment variables in tests
