### 2024-06-12 b233bce
chore: adjust type hint
