### 2024-06-13 42dbb8d
update cb implementation for subscriptions
Signed-off-by: iripiri <iris.m.koester@gmail.com>
