### 2024-06-18 e50f00f
chore: remove the data points creation in test
