### 2024-06-18 7385261
chore: change version requirement for paho-mqtt in setup
