### 2024-06-18 fdb468e
fix: adjust f strings to python 3.9
