### 2024-06-24 251879e
fixes to entity tests and implementation after review
Signed-off-by: iripiri <iris.m.koester@gmail.com>
