### 2024-06-25 8d60103
docs: update docs for requirements.txt
