### 2024-06-25 204b592
feat: implement alterationtypes in data model
