### 2024-06-25 f5739e2
docs: update CHANGELOG.md
