### 2024-06-25 5cbdf40
chore: use environment variables from settings
