### 2024-06-25 34a7223
chore: use environment variables from settings
