### 2024-06-26 4984cbf
fix: remove unset fields in batch operations
