### 2024-06-26 43c9b44
fix: remove unset fields in batch operations
