### 2024-06-27 fbdbd2b
Adding new attributes to the HttpCustom model (json, ngsi and timeout). Also added some tests for the pydantic model.
