+++ tests/models/test_ngsi_v2_subscriptions.py
@@ -90,6 +90,10 @@ class TestSubscriptions(unittest.TestCase):
         with self.assertRaises(ValidationError):
             Mqtt(url="mqtt://test.de:1883",
                  topic='/,t')
+        with self.assertRaises(ValidationError):
+            HttpCustom(url="https://working-url.de:80", json={}, ngsi={})
+        with self.assertRaises(ValidationError):
+            HttpCustom(url="https://working-url.de:80", payload="", json={})
         httpCustom = HttpCustom(url=self.http_url)
         mqtt = Mqtt(url=self.mqtt_url,
                     topic=self.mqtt_topic)
@@ -100,10 +104,18 @@ class TestSubscriptions(unittest.TestCase):
         notification = Notification.model_validate(self.notification)
         with self.assertRaises(ValidationError):
             notification.mqtt = httpCustom
+        notification = Notification.model_validate(self.notification)
         with self.assertRaises(ValidationError):
             notification.mqtt = mqtt
+        notification = Notification.model_validate(self.notification)
         with self.assertRaises(ValidationError):
             notification.mqtt = mqttCustom
+        notification = Notification.model_validate(self.notification)
+        with self.assertRaises(ValidationError):
+            notification.httpCustom = httpCustom
+        notification = Notification.model_validate(self.notification)
+        with self.assertRaises(ValidationError):
+            notification.exceptAttrs = ["temperature", "humidity"]
 
         # test onlyChangedAttrs-field
         notification = Notification.model_validate(self.notification)
