+++ filip/models/ngsi_v2/subscriptions.py
@@ -14,6 +14,11 @@ from .base import AttrsFormat, EntityPattern, Http, Status, Expression
 from filip.utils.validators import validate_mqtt_url, validate_mqtt_topic
 from filip.models.ngsi_v2.context import ContextEntity
 from filip.custom_types import AnyMqttUrl
+import warnings
+
+# The pydantic models still have a .json() function, but this method is deprecated.
+warnings.filterwarnings("ignore", category=UserWarning,
+                        message='Field name "json" shadows an attribute in parent "Http"')
 
 
 class Message(BaseModel):
@@ -69,6 +74,32 @@ class HttpCustom(Http):
                     'default payload (see "Notification Messages" sections) '
                     'is used.'
     )
+    json: Optional[Dict[str, Union[str, Json]]] = Field(
+        default=None,
+        description="JSON-based payload to be used in notifications. See 'JSON Payloads' section for more details."
+    )
+    ngsi: Optional[Dict[str, Union[str, Json]]] = Field(
+        default=None,
+        description="NGSI patching for payload to be used in notifications. See 'NGSI payload patching' section for "
+                    "more details."
+    )
+    timeout: Optional[int] = Field(
+        default=None,
+        description="Maximum time (in milliseconds) the subscription waits for the response. The maximum value allowed "
+                    "for this parameter is 1800000 (30 minutes). If timeout is defined to 0 or omitted, then the value "
+                    "passed as -httpTimeout CLI parameter is used. See section in the 'Command line options' for more "
+                    "details."
+    )
+
+    @model_validator(mode='after')
+    def validate_http(self):
+        fields = [self.payload, self.json, self.ngsi]
+        filled_fields = [field for field in fields if field is not None]
+
+        if len(filled_fields) > 1:
+            raise ValueError("Only one of payload, json or ngsi fields accepted at the same time in httpCustom.")
+
+        return self
 
 
 class Mqtt(BaseModel):
@@ -82,7 +113,7 @@ class Mqtt(BaseModel):
                     'only includes host and port)')
     topic: str = Field(
         description='to specify the MQTT topic to use',
-        )
+    )
     valid_type = field_validator("topic")(validate_mqtt_topic)
     qos: Optional[int] = Field(
         default=0,
@@ -205,17 +236,17 @@ class Notification(BaseModel):
                     '[A=0, B=null, C=null]. This '
     )
 
-    @field_validator('httpCustom')
-    def validate_http(cls, http_custom, values):
-        if http_custom is not None:
-            assert values['http'] is None
-        return http_custom
+    @model_validator(mode='after')
+    def validate_http(self):
+        if self.httpCustom is not None:
+            assert self.http is None
+        return self
 
-    @field_validator('exceptAttrs')
-    def validate_attr(cls, except_attrs, values):
-        if except_attrs is not None:
-            assert values['attrs'] is None
-        return except_attrs
+    @model_validator(mode='after')
+    def validate_attr(self):
+        if self.exceptAttrs is not None:
+            assert self.attrs is None
+        return self
 
     @model_validator(mode='after')
     def validate_endpoints(self):
@@ -247,7 +278,7 @@ class Response(Notification):
                     'Last notification timestamp in ISO8601 format.'
     )
     lastFailure: Optional[datetime] = Field(
-        default = None,
+        default=None,
         description='(not editable, only present in GET operations): '
                     'Last failure timestamp in ISO8601 format. Not present if '
                     'subscription has never had a problem with notifications.'
@@ -342,21 +373,21 @@ class Subscription(BaseModel):
     )
     subject: Subject = Field(
         description="An object that describes the subject of the subscription.",
-        example={
+        examples=[{
             'entities': [{'idPattern': '.*', 'type': 'Room'}],
             'condition': {
[…]
