### 2024-06-27 10b9a48
Renaming notification payloads validation functions
