### 2024-06-27 c0ba701
Added ngsi and json notification payload types, to httpCustom and mqttCustom notifications
