### 2024-06-28 a858383
Added models for validation of ngsi patched payload type
