+++ filip/models/ngsi_v2/subscriptions.py
@@ -9,12 +9,64 @@ from pydantic import \
     field_validator, model_validator, ConfigDict, BaseModel, \
     conint, \
     Field, \
-    Json, \
-    AliasChoices
+    Json
 from .base import AttrsFormat, EntityPattern, Http, Status, Expression
-from filip.utils.validators import validate_mqtt_url, validate_mqtt_topic
+from filip.utils.validators import (
+    validate_mqtt_url,
+    validate_mqtt_topic
+)
 from filip.models.ngsi_v2.context import ContextEntity
+from filip.models.ngsi_v2.base import (
+    EntityPattern,
+    Expression,
+    BaseValueAttribute
+)
 from filip.custom_types import AnyMqttUrl
+
+
+class NotificationAttr(BaseValueAttribute):
+    """
+    Model for NGSI V2 type payload in httpCustom/mqttCustom notifications.
+    The difference between this model and the usual BaseValueAttribute model is that
+    a metadata field is not allowed.
+    In the absence of type/value in some attribute field, one should resort to partial 
+    representations ( as specified in the orion api manual), done by the BaseValueAttr.
+    model.
+    """
+    model_config=ConfigDict(extra="forbid")
+
+class NotificationEntity(BaseModel):
+    """
+    Model for NGSI V2 type payload in httpCustom/mqttCustom notifications.
+    Differences between this model and the usual Context entity models include:
+        - id and type are not mandatory
+        - a metadata Attribute field is not allowed
+    In the absence of type/value in some attribute field, one should resort to partial 
+    representations ( as specified in the orion api manual), done by the BaseValueAttr.
+    model.
+    """
+    model_config = ConfigDict(
+        extra="allow", validate_default=True
+    )
+    id: Optional[str] = Field(
+        default=None,
+        max_length=256,
+        min_length=1,
+        frozen=True
+    )
+    type: Optional[Union[str, Enum]] = Field(
+        default=None,
+        max_length=256,
+        min_length=1,
+        frozen=True,
+    )
+
+    @model_validator(mode='after')
+    def validate_notification_attrs(self):
+        for v in self.model_dump(exclude={"id","type"}).values():
+            assert isinstance(NotificationAttr.model_validate(v),NotificationAttr)
+        return self
+    
 
 
 class Message(BaseModel):
@@ -75,7 +127,7 @@ class HttpCustom(Http):
         description='get a json as notification. If omitted, the default'
                     'payload (see "Notification Messages" sections) is used.'
     )
-    ngsi:Optional[ContextEntity] = Field(
+    ngsi:Optional[NotificationEntity] = Field(
         default=None,
         description='get an NGSI-v2 normalized entity as notification.If omitted, '
                     'the default payload (see "Notification Messages" sections) is used.'
@@ -146,7 +198,7 @@ class MqttCustom(Mqtt):
         description='get a json as notification. If omitted, the default'
                     'payload (see "Notification Messages" sections) is used.'
     )
-    ngsi:Optional[ContextEntity] = Field(
+    ngsi:Optional[NotificationEntity] = Field(
         default=None,
         description='get an NGSI-v2 normalized entity as notification.If omitted, '
                     'the default payload (see "Notification Messages" sections) is used.'
