### 2024-07-01 ec471e4
Re-organize and expand subscription endpoint tests
Signed-off-by: iripiri <iris.m.koester@gmail.com>
