+++ tests/clients/test_ngsi_ld_subscription.py
@@ -25,6 +25,15 @@ class TestSubscriptions(unittest.TestCase):
     Test class for context broker models
     """
 
+    def cleanup(self):
+        """
+        Cleanup test subscriptions
+        """
+        sub_list = self.cb_client.get_subscription_list()
+        for sub in sub_list:
+            if sub.id.startswith('urn:ngsi-ld:Subscription:test_sub'):
+                self.cb_client.delete_subscription(sub.id)
+
     def setUp(self) -> None:
         """
         Setup test data
@@ -49,6 +58,42 @@ class TestSubscriptions(unittest.TestCase):
             "uri": "http://my.endpoint.org/notify",
             "accept": "application/json"
         })
+        self.mqtt_topic = ''.join([FIWARE_SERVICE, FIWARE_SERVICEPATH])
+        self.MQTT_BROKER_URL_INTERNAL = "mqtt://mosquitto:1883"
+        self.MQTT_BROKER_URL_EXPOSED = "mqtt://localhost:1883"
+        self.endpoint_mqtt = Endpoint(**{
+            "uri": "mqtt://my.host.org:1883/my/test/topic",
+            "accept": "application/json",  # TODO check whether it works
+        })
+        self.cb_client = ContextBrokerLDClient(url=settings.LD_CB_URL,
+                                               fiware_header=self.fiware_header)
+        self.endpoint_http = Endpoint(**{
+            "uri": "http://137.226.248.246:1027/ngsi-ld/v1/subscriptions",
+            "Content-Type": "application/json",
+            "Accept": "application/json"
+        }
+                                      )
+        self.cleanup()
+
+    def tearDown(self) -> None:
+        self.cleanup()
+
+    def test_post_subscription_http(self):
+        """
+        Create a new HTTP subscription.
+        Args:
+            - Request body: required
+        Returns:
+            - (201) successfully created subscription
+        Tests:
+            - Create a HTTP subscription and post it
+        """
+        attr_id = "attr"
+        id = "urn:ngsi-ld:Subscription:" + "test_sub0"
+        notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)
+        sub = Subscription(id=id, notification=notification_param, entities=[{"type": "Room"}])
+        self.cb_client.post_subscription(sub)
+
 
     def test_get_subscription_list(self):
         """
@@ -136,13 +181,13 @@ class TestSubscriptions(unittest.TestCase):
         #self.cb_client.post_subscription(sub)
         sub_list = self.cb_client.get_subscription_list()
         self.assertEqual(10, len(sub_list))
-        
+
         for sub in sub_post_list:
             self.assertIn(sub in sub_list)
-            
+
         for sub in sub_list:
             self.cb_client.delete_subscription(id=sub.id)
-            
+
      
         """Test 2"""
         for i in range(2): 
@@ -184,7 +229,22 @@ class TestSubscriptions(unittest.TestCase):
             - Create a subscription twice to one message and see if the message is 
                 received twice or just once.
         """
- 
+
+    def test_post_subscription_http_check_broker(self):
+        """
+        Create a new HTTP subscription.
+        Args:
+            - Request body: required
+        Returns:
+            - (201) successfully created subscription
+        Tests:
+            - Create a subscription and post something from this subscription
+                to see if the subscribed broker gets the message.
+            - Create a subscription twice to one message and see if the message is
+                received twice or just once.
+        """
+        pass
+
 
     def test_get_subscription(self):
         """
@@ -198,6 +258,133 @@ class TestSubscriptions(unittest.TestCase):
             - Subscribe to a message and see if it appears when the message is subscribed to
[…]
