+++ tests/clients/test_ngsi_ld_subscription.py
@@ -16,7 +16,7 @@ from filip.models.ngsi_ld.subscriptions import \
     NotificationParams, \
     Subscription
 from filip.utils.cleanup import clear_all, clean_test
-from tests.config import TestSettings
+from tests.config import settings
 from random import randint
 
 
@@ -24,6 +24,15 @@ class TestSubscriptions(unittest.TestCase):
     """
     Test class for context broker models
     """
+
+    def cleanup(self):
+        """
+        Cleanup test subscriptions
+        """
+        sub_list = self.cb_client.get_subscription_list()
+        for sub in sub_list:
+            if sub.id.startswith('urn:ngsi-ld:Subscription:test_sub'):
+                self.cb_client.delete_subscription(sub.id)
 
     def setUp(self) -> None:
         """
@@ -54,137 +63,39 @@ class TestSubscriptions(unittest.TestCase):
             "uri": "mqtt://my.host.org:1883/my/test/topic",
             "accept": "application/json",  # TODO check whether it works
         })
-        CB_URL = "http://137.226.248.246:1027"
-        self.cb_client = ContextBrokerLDClient(url=CB_URL, fiware_header=self.fiware_header)        
+        self.cb_client = ContextBrokerLDClient(url=settings.LD_CB_URL, fiware_header=self.fiware_header)
         self.endpoint_http = Endpoint(**{
             "uri": "http://137.226.248.246:1027/ngsi-ld/v1/subscriptions",
             "Content-Type": "application/json",
             "Accept": "application/json"
         }
      )
+        self.cleanup()
 
-    def test_get_subscription_list(self):
+    def tearDown(self) -> None:
+        self.cleanup()
+
+
+    def test_post_subscription_http(self):
         """
-        Get a list of all current subscriptions the broker has subscribed to.
-        Args: 
-            - limit(number($double)): Limits the number of subscriptions retrieved
+        Create a new HTTP subscription.
+        Args:
+            - Request body: required
         Returns:
-            - (200) list of subscriptions
-        Tests for get subscription list:
-            - Get the list of subscriptions and get the count of the subsciptions -> compare the count
-            - Go through the list and have a look at duplicate subscriptions
-            - Set a limit for the subscription number and compare the count of subscriptions sent with the limit
-            - Save beforehand all posted subscriptions and see if all the subscriptions exist in the list -> added to Test 1
+            - (201) successfully created subscription
+        Tests:
+            - Create a HTTP subscription and post it
         """
-        
-        """Test 1"""
-        sub_post_list = list()
-        for i in range(10): 
-            attr_id = "attr" + str(i)
-            attr = {attr_id: ContextProperty(value=randint(0,50))}
-            id = "test_sub" + str(i)
-            uri_string =  "mqtt://my.host.org:1883/topic/" + str(i)
-            
-            endpoint_mqtt = Endpoint(**{
-                "uri": uri_string,
-                "accept": "application/json", 
-                "notifierInfo": [
-                    {
-                        "key": "MQTT-Version",
-                        "value": "mqtt5.0"
-                    }
-                ]
-            })
-            notification_param = NotificationParams(attributes=[attr_id], endpoint=endpoint_mqtt)  
-            sub = Subscription(id=id, notification=notification_param)
-            self.cb_client.post_subscription(sub)
-        # attr_id = "attr" + str(1)
-        # attr = {attr_id: ContextProperty(value=randint(0,50))}
-        # id = "test_sub" + str(1)
-        # uri_string =  "mqtt://my.host.org:1883/topic/" + str(1)
-        sub_example = {
-        "description": "Subscription to receive MQTT-Notifications about "
-                       "urn:ngsi-ld:Room:001",
-        "subject": {
-            "entities": [
-                {
-                    "id": "urn:ngsi-ld:Room:001",
[…]
