### 2024-07-02 0c198ab
docs: adapt doc strings for some tests
