### 2024-07-02 573b5a6
Added test for comparaison of posted and retrieved subscription objects
