### 2024-07-03 #6538 @abhishektiwari
@srmnitc Happy to be reviewer on this. Looks straightforward Python package for IoT APIs.
