### 2024-07-03 35018c3
docs: adapt doc strings length
