### 2024-07-03 #6538 @srmnitc
> @srmnitc Happy to be reviewer on this. Looks straightforward Python package for IoT APIs.

Thanks a lot!
