### 2024-07-03 #6953 @editorialbot
[REVIEW]: FiLiP: A python software development kit (SDK) for accelerating the development of services based on FIWARE IoT platform
**Submitting author:** @djs0109 (<a href="http://orcid.org/0000-0003-2247-2423">Junsong Du</a>)
**Repository:** https://github.com/RWTH-EBC/FiLiP
**Branch with paper.md** (empty if default branch): JOSS_submission
**Version:** v0.5.0
