+++ tests/clients/test_ngsi_ld_subscription.py
@@ -95,7 +95,7 @@ class TestSubscriptions(unittest.TestCase):
 
     def test_post_subscription_http_check_broker(self):
         """
-        Create a new HTTP subscription.
+        Create a new HTTP subscription and check whether messages are received.
         Args:
             - Request body: required
         Returns:
@@ -117,8 +117,7 @@ class TestSubscriptions(unittest.TestCase):
             - (200) subscription or empty list if successful 
             - Error Code
         Tests:
-            - Subscribe to a message and see if it appears when the message is subscribed to
-            - Choose a non-existent ID and see if the return is an empty array
+            - Get Subscription and check if the subscription is the same as the one posted
         """
         attr_id = "attr"
         id = "urn:ngsi-ld:Subscription:" + "test_sub0"
@@ -202,20 +201,6 @@ class TestSubscriptions(unittest.TestCase):
 #        #self.cb_client.post_subscription(sub)
         pass
 
-    def test_subscription_check_notifications(self):
-        """
-        Create a new MQTT subscription and check if messages are received.
-        Args:
-            - Request body: required
-        Returns:
-            - (201) successfully created subscription
-        Tests:
-            - Create a subscription and post something from this subscription
-                to see if the subscribed broker gets the message.
-            - Create a subscription twice to one message and see if the message is
-                received twice or just once.
-        """
-        pass
 
     def test_get_subscription_list(self):
         """
@@ -225,9 +210,8 @@ class TestSubscriptions(unittest.TestCase):
         Returns:
             - (200) list of subscriptions
         Tests for get subscription list:
-            - Get the list of subscriptions and get the count of the subsciptions -> compare the count
-            - Go through the list and have a look at duplicate subscriptions
-            - Set a limit for the subscription number and compare the count of subscriptions sent with the limit
+            - Create list of subscriptions and get the list of subscriptions -> compare the lists
+            - Set a limit for the subscription number and compare the count of subscriptions
         """
         sub_post_list = list()
         for i in range(10):
@@ -266,15 +250,22 @@ class TestSubscriptions(unittest.TestCase):
             sub = Subscription(id=id, notification=notification_param, entities=[{"type": "Room"}])
 
             if i == 0: 
-                subscription = sub
+                del_sub = sub
+                del_id = id
             self.cb_client.post_subscription(sub)
         
-        self.cb_client.delete_subscription(subscription_id=id)
         sub_list = self.cb_client.get_subscription_list(limit=10)
-        self.assertNotIn(subscription, sub_list)
+        sub_id_list = [sub.id for sub in sub_list]
+        self.assertIn(del_sub.id, sub_id_list)
+
+        self.cb_client.delete_subscription(subscription_id=del_id)
+        sub_list = self.cb_client.get_subscription_list(limit=10)
+        sub_id_list = [sub.id for sub in sub_list]
+        self.assertNotIn(del_sub.id, sub_id_list)
 
         for sub in sub_list:
             self.cb_client.delete_subscription(subscription_id=sub.id)
+
 
     def test_delete_subscription_check_broker(self):
         """
@@ -291,7 +282,8 @@ class TestSubscriptions(unittest.TestCase):
             
     def test_update_subscription(self):
         """
-        Only the fileds included in the request are updated in the subscription. 
+        Update a subscription.
+        Only the fields included in the request are updated in the subscription.
         Args:
             - subscriptionID(string): required
             - body(body): required
@@ -299,7 +291,7 @@ class TestSubscriptions(unittest.TestCase):
             - Successful: 204, no content
         Tests: 
             - Patch existing subscription and read out if the subscription got patched.
-            - Try to patch non-existent subscriüptions.
+            - Try to patch non-existent subscriptions.
[…]
