### 2024-07-03 aedd162
cleanup delete test and test descriptions
Signed-off-by: iripiri <iris.m.koester@gmail.com>
