+++ tests/clients/test_ngsi_ld_subscription.py
@@ -98,7 +98,7 @@ class TestSubscriptions(unittest.TestCase):
     def test_get_subscription_list(self):
         """
         Get a list of all current subscriptions the broker has subscribed to.
-        Args: 
+        Args:
             - limit(number($double)): Limits the number of subscriptions retrieved
             - offset(number($double)): Skip a number of subscriptions
             - options(string): Options dictionary("count")
@@ -111,10 +111,10 @@ class TestSubscriptions(unittest.TestCase):
             - Set offset for the subscription to retrive and check if the offset was procceded correctly.
             - Save beforehand all posted subscriptions and see if all the subscriptions exist in the list -> added to Test 1
         """
-        
+
         """Test 1"""
         sub_post_list = list()
-        for i in range(10): 
+        for i in range(10):
             attr_id = "attr" + str(i)
             attr = {attr_id: ContextProperty(value=randint(0,50))}
             id = "test_sub" + str(i)
@@ -188,26 +188,26 @@ class TestSubscriptions(unittest.TestCase):
         for sub in sub_list:
             self.cb_client.delete_subscription(id=sub.id)
 
-     
+
         """Test 2"""
-        for i in range(2): 
+        for i in range(2):
             attr_id = "attr"
             attr = {attr_id: ContextProperty(value=20)}
-            notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)       
-            id = "test_sub" 
+            notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)
+            id = "test_sub"
             sub = Subscription(id=id, notification=notification_param)
             self.cb_client.post_subscription(sub)
         sub_list = self.cb_client.get_subscription_list()
         self.assertNotEqual(sub_list[0], sub_list[1])
         for sub in sub_list:
             self.cb_client.delete_subscription(id=sub.id)
-            
-            
+
+
         """Test 3"""
-        for i in range(10): 
+        for i in range(10):
             attr_id = "attr" + str(i)
             attr = {attr_id: ContextProperty(value=randint(0,50))}
-            notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)       
+            notification_param = NotificationParams(attributes=[attr_id], endpoint=self.endpoint_http)
             id = "test_sub" + str(i)
             sub = Subscription(id=id, notification=notification_param)
             self.cb_client.post_subscription(sub)
@@ -215,24 +215,24 @@ class TestSubscriptions(unittest.TestCase):
         self.assertEqual(5, len(sub_list))
         for sub in sub_list:
             self.cb_client.delete_subscription(id=sub.id)
-            
+
     def test_post_subscription(self):
         """
         Create a new subscription.
         Args:
             - Request body: required
         Returns:
-            - (201) successfully created subscription 
+            - (201) successfully created subscription
         Tests:
-            - Create a subscription and post something from this subscription 
+            - Create a subscription and post something from this subscription
                 to see if the subscribed broker gets the message.
-            - Create a subscription twice to one message and see if the message is 
+            - Create a subscription twice to one message and see if the message is
                 received twice or just once.
         """
 
     def test_post_subscription_http_check_broker(self):
         """
-        Create a new HTTP subscription.
+        Create a new HTTP subscription and check whether messages are received.
         Args:
             - Request body: required
         Returns:
@@ -245,7 +245,6 @@ class TestSubscriptions(unittest.TestCase):
         """
         pass
 
-
     def test_get_subscription(self):
         """
         Returns the subscription if it exists.
@@ -255,8 +254,7 @@ class TestSubscriptions(unittest.TestCase):
             - (200) subscription or empty list if successful 
             - Error Code
         Tests:
[…]
