### 2024-07-03 8001b3a
feat: adjust path finding for .env.filip
