### 2024-07-08 527ca29
docs: Update README.md for local development
