### 2024-07-09 5c64d58
chore: fix the position of .env.filip
