### 2024-07-09 47db06f
chore: use settings in examples
