+++ examples/ngsi_v2/e08_ngsi_v2_iota_paho_mqtt.py
@@ -10,7 +10,7 @@ import time
 import paho.mqtt.client as mqtt
 
 from urllib.parse import urlparse
-
+from filip.config import settings
 from filip.models import FiwareHeader
 from filip.models.ngsi_v2.iot import \
     Device, \
@@ -28,11 +28,11 @@ from filip.clients.ngsi_v2 import HttpClient, HttpClientConfig
 # addresses:
 #
 # Host address of Context Broker
-CB_URL = "http://localhost:1026"
+CB_URL = settings.CB_URL
 # Host address of IoT-Agent
-IOTA_URL = "http://localhost:4041"
+IOTA_URL = settings.IOTA_URL
 # Host address of the MQTT-Broker
-MQTT_BROKER_URL = "mqtt://localhost:1883"
+MQTT_BROKER_URL = settings.MQTT_BROKER_URL
 
 # Here you can also change the used Fiware service
 # FIWARE-Service
