+++ examples/ngsi_v2/e03_ngsi_v2_context_subscriptions_http.py
@@ -8,7 +8,7 @@
 import logging
 import datetime
 import time
-
+from filip.config import settings
 from filip.models.ngsi_v2.subscriptions import Subscription
 from filip.clients.ngsi_v2.cb import ContextBrokerClient
 from filip.models.base import FiwareHeader
@@ -19,7 +19,7 @@ from filip.models.base import FiwareHeader
 # You can set the address:
 #
 # Host address of Context Broker
-CB_URL = "http://localhost:1026"
+CB_URL = settings.CB_URL
 
 # You can also change the used Fiware service
 # FIWARE-Service
