+++ examples/ngsi_v2/e12_ngsi_v2_use_case_models.py
@@ -14,9 +14,9 @@ from filip.models.ngsi_v2.context import ContextEntityKeyValues
 from filip.clients.ngsi_v2.cb import ContextBrokerClient
 from filip.utils.cleanup import clear_context_broker
 from pprint import pprint
-
+from filip.config import settings
 # Host address of Context Broker
-CB_URL = "http://localhost:1026"
+CB_URL = settings.CB_URL
 
 # You can here also change the used Fiware service
 # FIWARE-Service
