+++ examples/ngsi_v2/e01_ngsi_v2_context_basics.py
@@ -10,6 +10,7 @@ from filip.models.base import FiwareHeader, DataType
 from filip.models.ngsi_v2.context import ContextEntity, ContextAttribute, \
     NamedContextAttribute
 from filip.utils.simple_ql import QueryString
+from filip.config import settings
 
 # ## Parameters
 #
@@ -17,7 +18,7 @@ from filip.utils.simple_ql import QueryString
 # You can set the address here:
 #
 # Host address of Context Broker
-CB_URL = "http://localhost:1026"
+CB_URL = settings.CB_URL
 
 # You can also change the used Fiware service
 # FIWARE-Service
