### 2024-07-09 ded1f43
chore: convert mqtt url to str
