### 2024-07-09 a24273c
docs: add hint regarding .env.filip in readme
