### 2024-07-10 3b8b977
fix: iota_url check seperately checked now.
