### 2024-07-10 317cfb5
chore: add a new test entity to test the nested structure of properties
