### 2024-07-11 6be7469
Added verification for the test_post_subscription_http test
