### 2024-07-11 c91e8b7
Added verification for the test_post_subscription_http test
