### 2024-07-11 f58482c
Added validation for Content-Type header
