### 2024-07-12 e39289f
Added subscription test for mqtt endpoint
