### 2024-07-12 252a5f8
Added subscription test for mqtt endpoint
