### 2024-07-13 RWTH-EBC/FiLiP#303 @abhishektiwari
[JOSS Review] Broken functionality documentation link
Documentation [link](https://ebc.pages.rwth-aachen.de/EBC_all/github_ci/FiLiP/master/docs/index.html) is broken in `README.md`. Please fix it and provide working link so I can review the docs.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6953)
