### 2024-07-13 #6953 @abhishektiwari
HI @djs0109,
Please see my initial comments,

1.  Please fix/provide client documentation link as [requested here](https://github.com/RWTH-EBC/FiLiP/issues/303) so I can validate functional claims of the software.
2. Software paper is light on the state of the field. There seems to be multiple independent Python clients for FIWARE  (e.g.  [1](https://github.com/OkinawaOpenLaboratory/QuantumLeapClient), [2](https://github.com/telefonicaid/fiware-orion), [3](https://github.com/robotics-4-all/fiware-ngsi-api), [4](https://github.com/FIoT-Client/fiot-client-ngsi-python), etc.) but none of have been mentioned in the paper. These clients along with automatically generated API clients using OpenAPI specification should be compared in the paper. FiLiP provides features like strong client side validation, domain models, json-schema format due to its use of Pydantic package. Any of other FIWARE clients can use Pydantic to provide these features. So I am not sure why paper is over stressing these points when they are basically inherited from a underlaying dependency.
3. For a software paper it is quite a long read. Consider trimming some of the sections like conclusion which is unnecessary too long.
