### 2024-07-15 RWTH-EBC/FiLiP#303 @djs0109
Need to check settings in gitlab-ci.yml:

- [ ] COVERAGE_TYPE
- [ ] EXAMPLE_CONVERTER_CONFIG
- [ ] template links for gitlab-ci pages
- [ ] template links for coverage
- [ ] Update links in readme.md
