### 2024-07-15 a6a33bc
Update 'docs' folder with 303--JOSS-Review-Broken-functionality-documentation-link and delete old branches.
