+++ docs/303--JOSS-Review-Broken-functionality-documentation-link/docs/_sources/code/filip.clients.rst.txt
@@ -1,0 +1,28 @@
+filip.clients package
+=====================
+
+.. automodule:: filip.clients
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   filip.clients.mqtt
+   filip.clients.ngsi_ld
+   filip.clients.ngsi_v2
+
+Submodules
+----------
+
+filip.clients.base\_http\_client module
+---------------------------------------
+
+.. automodule:: filip.clients.base_http_client
+   :members:
+   :undoc-members:
+   :show-inheritance:
