+++ docs/303--JOSS-Review-Broken-functionality-documentation-link/docs/_sources/code/filip.clients.ngsi_ld.rst.txt
@@ -1,0 +1,7 @@
+filip.clients.ngsi\_ld package
+==============================
+
+.. automodule:: filip.clients.ngsi_ld
+   :members:
+   :undoc-members:
+   :show-inheritance:
