### 2024-07-15 85abe6f
test: Adding tests for the mqtt custom notifications, that use payload, json and ngsi messages
