+++ tests/clients/test_ngsi_v2_cb.py
@@ -910,7 +910,91 @@ class TestContextBroker(unittest.TestCase):
             "throttling": 0
         })
 
+        mqtt_custom_topic = "notification/custom"
+        sub_with_mqtt_custom_notification_payload = Subscription.model_validate({
+            "description": "Test mqtt custom notification with payload message",
+            "subject": {
+                "entities": [
+                    {
+                        "id": "Test:001",
+                        "type": "Test"
+                    }
+                ]
+            },
+            "notification": {
+                "mqttCustom": {
+                    "url": mqtt_url_internal,
+                    "topic": mqtt_custom_topic,
+                    "payload": "The value of the %22temperature%22 attribute %28of the device ${id}, ${type}%29 is"
+                               " ${temperature}. Humidity is ${humidity} and CO2 is ${co2}."
+                },
+                "attrs": ["temperature", "humidity", "co2"],
+                "onlyChangedAttrs": False
+            },
+            "expires": datetime.now() + timedelta(days=1),
+            "throttling": 0
+        })
+
+        sub_with_mqtt_custom_notification_json = Subscription.model_validate({
+            "description": "Test mqtt custom notification with json message",
+            "subject": {
+                "entities": [
+                    {
+                        "id": "Test:001",
+                        "type": "Test"
+                    }
+                ]
+            },
+            "notification": {
+                "mqttCustom": {
+                    "url": mqtt_url_internal,
+                    "topic": mqtt_custom_topic,
+                    "json": {
+                        "t": "${temperature}",
+                        "h": "${humidity}",
+                        "c": "${co2}"
+                    }
+                },
+                "attrs": ["temperature", "humidity", "co2"],
+                "onlyChangedAttrs": False
+            },
+            "expires": datetime.now() + timedelta(days=1),
+            "throttling": 0
+        })
+
+        sub_with_mqtt_custom_notification_ngsi = Subscription.model_validate({
+            "description": "Test mqtt custom notification with ngsi message",
+            "subject": {
+                "entities": [
+                    {
+                        "id": "Test:001",
+                        "type": "Test"
+                    }
+                ]
+            },
+            "notification": {
+                "mqttCustom": {
+                    "url": mqtt_url_internal,
+                    "topic": mqtt_custom_topic,
+                    "ngsi": {
+                        "id": "prefix:${id}",
+                        "type": "newType",
+                        "temperature": {
+                            "value": 123,
+                            "type": "Number"
+                        },
+
+                    }
+                },
+                "attrs": ["temperature", "humidity", "co2"],
+                "onlyChangedAttrs": False
+            },
+            "expires": datetime.now() + timedelta(days=1),
+            "throttling": 0
+        })
+
         # MQTT settings
+        custom_sub_message = None
         sub_message = None
         sub_messages = {}
 
@@ -923,6 +1007,7 @@ class TestContextBroker(unittest.TestCase):
                 logger.info("Successfully, connected with result code " + str(
                     reasonCode))
             client.subscribe(mqtt_topic)
+            client.subscribe(mqtt_custom_topic)
 
         def on_subscribe(client, userdata, mid, granted_qos, properties=None):
             logger.info("Successfully subscribed to with QoS: %s", granted_qos)
@@ -931,8 +1016,12 @@ class TestContextBroker(unittest.TestCase):
             logger.info("Received MQTT message: " + msg.topic + " " + str(
                 msg.payload))
             nonlocal sub_message
-            sub_message = Message.model_validate_json(msg.payload)
-            sub_messages[sub_message.subscriptionId] = sub_message
+            nonlocal custom_sub_message
[…]
