+++ docs/source/conf.py
@@ -10,6 +10,7 @@
 import os
 import sys
 import datetime
+from pathlib import Path
 # pylint: disable-all
 
 sys.path.insert(0, os.path.abspath('../..'))
@@ -24,7 +25,15 @@ copyright = f'2021-{datetime.datetime.now().year}, RWTH Aachen University, ' \
 author = 'E.ON ERC - EBC'
 
 # The full version, including alpha/beta/rc tags
-release = '0.2.4'
+# Get the version from FiliP/filip/__init__.py:
+with open(Path(__file__).parents[2].joinpath(project,
+                                             "filip", "__init__.py"), "r") as file:
+    for line in file.readlines():
+        if line.startswith("__version__"):
+            release = line.replace("__version__", "").split("=")[1].strip().replace(
+                "'", "").replace(
+                '"', '')
+release = release
 
 # The short X.Y version.
 version = '.'.join(release.split('.')[0:2])
@@ -146,7 +155,7 @@ html_context = {
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = []
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
