### 2024-07-15 #6953 @djs0109
Hi @abhishektiwari, thank you for the comments! We are already working on the docs issue. Hi @srmnitc , should we update the `paper.md` directly in the submission branch after revision?
