### 2024-07-15 bf1f5b0
Update 'docs' folder with master and delete old branches.
