### 2024-07-15 2c2dfdb
Update 'docs' folder with 306-example-is-not-correctly-linked-in-documentation and delete old branches.
