+++ docs/306-example-is-not-correctly-linked-in-documentation/docs/_sources/code/filip.models.ngsi_ld.rst.txt
@@ -1,0 +1,7 @@
+filip.models.ngsi\_ld package
+=============================
+
+.. automodule:: filip.models.ngsi_ld
+   :members:
+   :undoc-members:
+   :show-inheritance:
