### 2024-07-15 72cabe7
Update 'docs' folder with JOSS_submission and delete old branches.
