+++ docs/JOSS_submission/docs/_sources/api/filip.semantics.ontology_parser.rst.txt
@@ -1,0 +1,37 @@
+filip.semantics.ontology\_parser package
+========================================
+
+Submodules
+----------
+
+filip.semantics.ontology\_parser.post\_processer module
+-------------------------------------------------------
+
+.. automodule:: filip.semantics.ontology_parser.post_processer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.semantics.ontology\_parser.rdfparser module
+-------------------------------------------------
+
+.. automodule:: filip.semantics.ontology_parser.rdfparser
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.semantics.ontology\_parser.vocabulary\_builder module
+-----------------------------------------------------------
+
+.. automodule:: filip.semantics.ontology_parser.vocabulary_builder
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: filip.semantics.ontology_parser
+   :members:
+   :undoc-members:
+   :show-inheritance:
