+++ docs/JOSS_submission/docs/_sources/examples.rst.txt
@@ -1,0 +1,1 @@
+.. mdinclude::  ../../examples/README.sphinx.md
