+++ docs/JOSS_submission/docs/_sources/api/filip.semantics.rst.txt
@@ -1,0 +1,46 @@
+filip.semantics package
+=======================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   filip.semantics.ontology_parser
+   filip.semantics.vocabulary
+
+Submodules
+----------
+
+filip.semantics.semantics\_manager module
+-----------------------------------------
+
+.. automodule:: filip.semantics.semantics_manager
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.semantics.semantics\_models module
+----------------------------------------
+
+.. automodule:: filip.semantics.semantics_models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.semantics.vocabulary\_configurator module
+-----------------------------------------------
+
+.. automodule:: filip.semantics.vocabulary_configurator
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: filip.semantics
+   :members:
+   :undoc-members:
+   :show-inheritance:
