+++ docs/JOSS_submission/docs/_sources/api/filip.models.ngsi_v2.rst.txt
@@ -1,0 +1,69 @@
+filip.models.ngsi\_v2 package
+=============================
+
+Submodules
+----------
+
+filip.models.ngsi\_v2.base module
+---------------------------------
+
+.. automodule:: filip.models.ngsi_v2.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.models.ngsi\_v2.context module
+------------------------------------
+
+.. automodule:: filip.models.ngsi_v2.context
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.models.ngsi\_v2.iot module
+--------------------------------
+
+.. automodule:: filip.models.ngsi_v2.iot
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.models.ngsi\_v2.registrations module
+------------------------------------------
+
+.. automodule:: filip.models.ngsi_v2.registrations
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.models.ngsi\_v2.subscriptions module
+------------------------------------------
+
+.. automodule:: filip.models.ngsi_v2.subscriptions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.models.ngsi\_v2.timeseries module
+---------------------------------------
+
+.. automodule:: filip.models.ngsi_v2.timeseries
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.models.ngsi\_v2.units module
+----------------------------------
+
+.. automodule:: filip.models.ngsi_v2.units
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: filip.models.ngsi_v2
+   :members:
+   :undoc-members:
+   :show-inheritance:
