+++ docs/JOSS_submission/docs/_sources/api/filip.clients.rst.txt
@@ -1,0 +1,31 @@
+filip.clients package
+=====================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   filip.clients.mqtt
+   filip.clients.ngsi_ld
+   filip.clients.ngsi_v2
+
+Submodules
+----------
+
+filip.clients.base\_http\_client module
+---------------------------------------
+
+.. automodule:: filip.clients.base_http_client
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: filip.clients
+   :members:
+   :undoc-members:
+   :show-inheritance:
