### 2024-07-16 #6953 @srmnitc
> Hi @abhishektiwari, thank you for the comments! We are already working on the docs issue. Hi @srmnitc , should we update the `paper.md` directly in the submission branch after revision?

Yes, you can update it there.
