### 2024-07-17 5e69c21
fix: iota error in clean all function
