### 2024-07-17 b2355ca
chore: deactivate the pages deployment for other branches
