+++ docs/283-clear_all-function-error-if-iota_url-is-none/docs/_sources/code/filip.semantics.vocabulary.rst.txt
@@ -1,0 +1,50 @@
+filip.semantics.vocabulary package
+==================================
+
+.. automodule:: filip.semantics.vocabulary
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+filip.semantics.vocabulary.combined\_relations module
+-----------------------------------------------------
+
+.. automodule:: filip.semantics.vocabulary.combined_relations
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.semantics.vocabulary.entities module
+------------------------------------------
+
+.. automodule:: filip.semantics.vocabulary.entities
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.semantics.vocabulary.relation module
+------------------------------------------
+
+.. automodule:: filip.semantics.vocabulary.relation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.semantics.vocabulary.source module
+----------------------------------------
+
+.. automodule:: filip.semantics.vocabulary.source
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+filip.semantics.vocabulary.vocabulary module
+--------------------------------------------
+
+.. automodule:: filip.semantics.vocabulary.vocabulary
+   :members:
+   :undoc-members:
+   :show-inheritance:
