### 2024-07-18 50c990d
chore: add hints about supported alteration types
