+++ joss/paper.bib
@@ -46,6 +46,14 @@
 	langid = {english}
 }
 
+@misc{uncefact,
+	title = {Codes for Units of Measure Used in International Trade},
+	publisher = {UN Centre for Trade Facilitation},
+	journal = {UNECE Code List Recommendations},
+	url = {https://unece.org/trade/uncefact/cl-recommendations},
+	year = {2021}
+}
+
 @misc{pydantic,
   title = {Pydantic: Data validation using Python type hint},
   year = {2023},
@@ -61,6 +69,40 @@
   journal = {Python Documentation},
   url = {https://docs.python.org/3/library/unittest.html}
 }
+
+
+@misc{QuantumLeapClient,
+  title = {QLClient},
+  year = {2024},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/OkinawaOpenLaboratory/QuantumLeapClient}
+}
+
+@misc{fiware-ngsi-api,
+  title = {fiware-api},
+  year = {2024},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/robotics-4-all/fiware-ngsi-api}
+}
+
+@misc{fiot-client-ngsi-python,
+  title = {fiot-client},
+  year = {2024},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/FIoT-Client/fiot-client-ngsi-python}
+}
+
+@misc{orion-python-client,
+  title = {orion-client},
+  year = {2024},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/openath/orion-python-client}
+}
+
 
 @misc{requests,
   author = {Kenneth Reitz},
