+++ joss/paper.md
@@ -39,62 +39,64 @@ These components play a pivotal role in core context information management, IoT
 In this context, we developed FiLiP, a Python Software Development Kit (SDK) designed to accelerate the development of services that interact with the aforementioned FIWARE GEs. 
 FiLiP emerges as a powerful tool, particularly in the domain of energy management systems, where it simplifies development, enhances efficiency, and empowers developers to create reliable and reusable IoT applications, aligning with the digital transformation facilitated by FIWARE.
 
-# Motivation
+# Statement of need
 
 The NGSI-v2 specification defines a standardized RESTful API to represent, exchange and manage context information throughout its entire lifecycle [@canterafonsecaFIWARENGSIV2Specification].
 Researchers or application developers can refer to the specifications and build reliable and interoperable data systems.
 However, NGSI-v2 is a general cross-domain specification and does not provide any domain-specific information models. 
 Although it defines the general data structure of entities, the definition of domain-specific attributes and metadata information remains in the responsibility of the users.
 Hence, adopting the specification for domain-specific engineering applications requires extensive training in both data modeling and programming of IoT applications.
-Moreover, the specification has continuously evolved over time.
-Therefore, enormous efforts are required to, first, define domain specific data models and, second,
-to comply with the specifications and keep applications and models up-to-date with the latest version of the specification.
-Thus, although FIWARE provides OpenAPI specifications [@openapi] [@canterafonsecaFIWARENGSIV2Specification], which can be used to automatically generate API clients for various programming languages, there are still challenges that make those auto-generated API clients less reliable:
+Additionally, the specification has continuously evolved over time, necessitating significant efforts to define domain-specific data models and to ensure that applications and models comply with the latest specification.
+
+Existing FIWARE APIs client libraries can already provide some basic support but often fall short in several key areas.
+A comparison as shown in \autoref{Comparison} reveals the shortage of various existing libraries, including limited support for multiple FIWARE APIs, lack of advanced validation and data models integration, and unsatisfactory usability and reliability.
+
+| Library                    | API Support                                  | Validation         | Datamodel            | Learning Support                           | Tests                         |
+|----------------------------|----------------------------------------------|--------------------|----------------------|--------------------------------------------|-------------------------------|
+| [@QuantumLeapClient]       | TimeSeries                                   | -                  | -                    | Readme file                                | 18                            |
+| [@fiware-ngsi-api]         | Device, Entity                               | Basic    | Proprietary     | -                                          | -                             |
+| [@fiot-client-ngsi-python] | Device (incomplete), Entity, Subscription    | -                  | -                    | Readme file (documentation not deployed)   | 31                            |
+| [@orion-python-client]     | Entity, Subscription (incomplete)            | -                  | -                    | Readme file                                | 1                             |
+| FiLiP                      | Device, Entity, Subscription, TimeSeries     | Enhanced | Customizable         | Documentation, 12 examples, and 8 tutorials| 82 (84% coverage)             |
+\label{Comparison}
+
+Although FIWARE provides OpenAPI specifications [@openapi] [@canterafonsecaFIWARENGSIV2Specification], which can be used to automatically generate API clients for various programming languages, there are still issues that make those auto-generated API clients less reliable:
 
 - The quality of auto-generated code strongly depends on the provided input data, i.e. OpenAPI specification.
 - Integrating additional features is generally not viable because manipulating the generated code can result in larger maintenance efforts.
[…]
