### 2024-07-18 #6953 @djs0109
Hi @srmnitc @abhishektiwari , we have now adapted the paper based on your first comments. Here also a summary of what has been done:
1. we add a comparison of 4 existing libraries to support the state of need
2. Towards the concern in the **second point** of @abhishektiwari, we have the following "rebuttals": we can understand your concerns regarding the emphasis of the paper on features such as strong client-side validation, which are implemented based on the Pydantic package. It's true that these features are facilitated by Pydantic, but our implementation goes beyond merely utilizing these features. We have integrated Pydantic in a way that is tailored specifically to the FIWARE ecosystem. Our implementation extends the basic functionality, based on our practical experience, which can better suit the needs of FIWARE users. For instance, we have created a custom validator that validates the unit specified in metadata based on UN/CEFACT. The validator will enrich the metadata with further `symbol`, `conversion_factor`, and `description`, after validation runs through. These points are now also addressed and rephrased in the updated paper. I hope it would resolve some concerns here.
3. We have generally shortened the paper, especially in **implementation** and **conclusion**
