### 2024-07-21 #6953 @abhishektiwari
@djs0109 Is there a testing FIWARE account paper reviewers can use to fast track the validation of the functionality? I have requested the [testing account setup](https://cloud.lab.fiware.org/auth/login/) but I am not sure how long it will take for account to be created? Docker approach to setup FIWARE is quite involved and may not work on my Ubuntu Multipass running on my Mac.
