### 2024-07-21 #6953 @djs0109
@abhishektiwari Unfortunately, I don't know of any publicly available FWIARE test stack. But we have [here](https://github.com/N5GEH/n5geh.platform/tree/master/v2) a tutorial for setting up a basic FIWARE stack via docker. Some Mac users have good experience with it, so I hope it can also work for you. I am very willing to help if you encounter an unexpected issue.
