### 2024-07-23 #6953 @SBlechmann
> @abhishektiwari We run the unit-test on our own CI-runner, which can reach our internal FIWARE Stack. But unfortunately, I don't know of any publicly available FWIARE test stack. But we have [here](https://github.com/N5GEH/n5geh.platform/tree/master/v2) a tutorial for setting up a basic FIWARE stack via docker. Some Mac users have good experience with it, so I hope it can also work for you. I am very willing to help if you encounter an unexpected issue.

@abhishektiwari, please, also take into account possible requisited described in a parent folder of the linked tutorial: https://github.com/N5GEH/n5geh.platform/tree/master
If you have any questions regarding the setup of the platform, please, do not hesitate to reach out.
