### 2024-07-23 26ab40e
feat: add covered in notification model
