### 2024-07-23 56570f3
chore: test functional behavior of `covered`
