### 2024-07-23 796f569
fix: allow empty string as value for any data type
