### 2024-07-23 d7749b8
Put broker callback tests into their own class
