### 2024-07-23 33ddc8c
Put broker callback tests into their own class
