+++ tests/clients/test_ngsi_ld_subscription.py
@@ -36,10 +36,6 @@ class TestSubscriptions(unittest.TestCase):
         for sub in sub_list:
             if sub.id.startswith('urn:ngsi-ld:Subscription:test_sub'):
                 self.cb_client.delete_subscription(sub.id)
-        entity_list = self.cb_client.get_entity_list()
-        for entity in entity_list:
-            if entity.id.startswith('urn:ngsi-ld:Entity:test_entity'):
-                self.cb_client.delete_entity_by_id(entity_id=entity.id)
 
     def setUp(self) -> None:
         """
@@ -71,17 +67,6 @@ class TestSubscriptions(unittest.TestCase):
             "accept": "application/json",  # TODO check whether it works
         })
         self.cb_client = ContextBrokerLDClient(url=settings.LD_CB_URL, fiware_header=self.fiware_header)
-        self.mqtt_client = mqtt.Client(callback_api_version=CallbackAPIVersion.VERSION2)
-        #on_message callbacks differ from test to test, but on connect callbacks dont
-        def on_connect_fail(client,userdata):
-            self.fail("Test failed due to broker being down")
-
-        def on_connect(client,userdata,flags,reason_code,properties):
-            self.mqtt_client.subscribe("my/test/topic")
-
-        self.mqtt_client.on_connect_fail = on_connect_fail
-        self.mqtt_client.on_connect = on_connect
-
         self.endpoint_http = Endpoint(**{
             "uri": "http://137.226.248.246:1027/ngsi-ld/v1/subscriptions",
             "Content-Type": "application/json",
@@ -89,48 +74,6 @@ class TestSubscriptions(unittest.TestCase):
         }
      )
         self.cleanup()
-        self.entity_dict = {
-            'id':'urn:ngsi-ld:Entity:test_entity03',
-            'type':'Room',
-            'temperature':{
-                'type':'Property',
-                'value':30
-            }
-        }
-        #posting one single entity to check subscription existence/triggers
-        self.cb_client.post_entity(entity=ContextLDEntity(**self.entity_dict))
-
-        #copy and update a single dict as the corresponding test requires
-        self.sub_dict = {
-            'description':'Test Subscription',
-            'id':'urn:ngsi-ld:Subscription:test_sub25',
-            'type':'Subscription',
-            'entities':[
-                {
-                    'type':'Room'
-                }
-            ],
-            'watchedAttributes':[
-                'temperature'
-            ],
-            'q':'temperature<30',
-            'notification':{
-                'attributes':[
-                    'temperature'
-                ],
-                'format':'normalized',
-                'endpoint':{
-                    'uri':'mqtt://mosquitto:1883/my/test/topic', # change uri
-                    'Accept':'application/json'
-                },
-                'notifierInfo':[
-                    {
-                        "key":"MQTT-Version",
-                        "value":"mqtt5.0"
-                    }
-                ]
-            }
-        }
 
     def tearDown(self) -> None:
         self.cleanup()
@@ -190,54 +133,6 @@ class TestSubscriptions(unittest.TestCase):
         self.assertEqual(sub.entities, sub_get.entities)
         self.assertEqual(sub.notification.attributes, sub_get.notification.attributes)
         self.assertEqual(sub.notification.endpoint.uri, sub_get.notification.endpoint.uri)
-
-
-    def test_get_subscription_check_broker(self):
-        """
-        Returns the subscription if it exists.
-        Args:
-            - subscriptionId(string): required
-        Returns:
-            - (200) subscription or empty list if successful
-            - Error Code
-        Tests:
-            - Subscribe to a message and see if it appears when the message is subscribed to
-            - Choose a non-existent ID and see if the return is an empty array
-        """
-        pass
-
-    def test_post_subscription_mqtt(self):
-         #if a notification is not received before timer runs out, test is assumed failed
[…]
