### 2024-07-24 d7e88ee
Fixed missing quotes in Content-Type headers
