+++ tests/clients/test_ngsi_ld_subscription.py
@@ -366,45 +366,10 @@ class TestSubscriptions(unittest.TestCase):
         # TODO
 
 class TestSubsCheckBroker(unittest.TestCase):
-    entity_dict = {
-        'id':'urn:ngsi-ld:Entity:test_entity03',
-        'type':'Room',
-        'temperature':{
-            'type':'Property',
-            'value':30
-        }
-    }
-    
-    sub_dict = {
-        'description':'Test Subscription',
-        'id':'urn:ngsi-ld:Subscription:test_sub25',
-        'type':'Subscription',
-        'entities':[
-            {
-                'type':'Room'
-            }
-        ],
-        'watchedAttributes':[
-            'temperature'
-        ],
-        'q':'temperature<30',
-        'notification':{
-            'attributes':[
-                'temperature'
-            ],
-            'format':'normalized',
-            'endpoint':{
-                'uri':'mqtt://mosquitto:1883/my/test/topic', # change uri
-                'Accept':'application/json'
-            },
-            'notifierInfo':[
-                {
-                    "key":"MQTT-Version",
-                    "value":"mqtt5.0"
-                }
-            ]
-        }
-    }
+
+    @unittest.skip("Helper function for timer")
+    def timeout_func(x):
+            x[0] = False
 
     def cleanup(self):
         """
@@ -425,6 +390,45 @@ class TestSubsCheckBroker(unittest.TestCase):
         Returns:
             None
         """
+        self.entity_dict = {
+            'id':'urn:ngsi-ld:Entity:test_entity03',
+            'type':'Room',
+            'temperature':{
+                'type':'Property',
+                'value':30
+            }
+        }
+        
+        self.sub_dict = {
+            'description':'Test Subscription',
+            'id':'urn:ngsi-ld:Subscription:test_sub25',
+            'type':'Subscription',
+            'entities':[
+                {
+                    'type':'Room'
+                }
+            ],
+            'watchedAttributes':[
+                'temperature'
+            ],
+            'q':'temperature<30',
+            'notification':{
+                'attributes':[
+                    'temperature'
+                ],
+                'format':'normalized',
+                'endpoint':{
+                    'uri':'mqtt://mosquitto:1883/my/test/topic', # change uri
+                    'Accept':'application/json'
+                },
+                'notifierInfo':[
+                    {
+                        "key":"MQTT-Version",
+                        "value":"mqtt5.0"
+                    }
+                ]
+            }
+        }
         FIWARE_SERVICE = "service"
         FIWARE_SERVICEPATH = "/"
         self.fiware_header = FiwareLDHeader(
@@ -446,6 +450,15 @@ class TestSubsCheckBroker(unittest.TestCase):
         #posting one single entity to check subscription existence/triggers
         self.cb_client.post_entity(entity=ContextLDEntity(**self.entity_dict))
 
+        #All broker tests rely on awaiting a message. This timer helps with:
+        #   -Avoiding hang ups in the case of a lost connection
+        #   -Avoid ending the tests early, in the case a notification takes longer
+
+        self.timeout = 5 # in seconds
+        self.last_test_timeout = [True]
+        self.timeout_proc = threading.Timer(self.timeout,self.timeout_func,
+                                            args=[self.last_test_timeout])
+
 
     def tearDown(self) -> None:
         self.cleanup()
@@ -456,17 +469,9 @@ class TestSubsCheckBroker(unittest.TestCase):
         Tests:
             - Subscribe using an mqtt topic as endpoint and see if notification is received
         """
-        #Declare timer function, mqtt message callback and a check variable(test_res)
-        #Variable is in list because python threads get copies of primitive objects (e.g bool)
-        #but not of iterables
-        test_res = [True]
[…]
