### 2024-07-24 9a99d7f
Fixed mistake in parsing ContextProperty when updating an entity
