### 2024-07-24 dfc8722
test: update v2 relic keyword
