### 2024-07-24 15ae3bd
chore: use internal mqtt broker url in settings
