### 2024-07-25 #6953 @DiegoAscanio
@djs0109 There's a misreference to Table 1 at page 1 and there's no caption to explain the table.
