### 2024-07-25 #6953 @DiegoAscanio
I think that a straightforward example or a bundled a proof-of-concept (within a docker container and described on its respective section at the paper) would enhance your work's understandability as it would point out to any reader a basic problem that your library helps to solve in a better way than its competitors or than the FIWARE specifications out-of-the-box suggests to.
