### 2024-07-25 #6953 @DiegoAscanio
Although suggestions I made above about proof-of-concepts or out-of-the-box examples, the usage of FiLiP at literature are proofs that the functional claims of the software were already confirmed, as FiLiP was already cited in peer reviewed academic papers.

With a minor review regarding a misreference and a caption absence, both belonging to the first table at the paper, I therefore recommend this paper for publishing at JOSS.
