### 2024-07-26 #6953 @djs0109
Hi @DiegoAscanio , thank you for your thoughtful feedback!

The idea of including an interactive example within a Docker container is very inspiring, and we would like to implement this feature in the long term. Currently, as shown in the table, we have provided 12 [examples](https://github.com/RWTH-EBC/FiLiP/tree/master/examples) and 8 [tutorials](https://github.com/RWTH-EBC/FiLiP/tree/master/tutorials/ngsi_v2), most of which include a well-designed use case or storyline. Additionally, we have described an intuitive example in the `Use Case` section of this paper to further illustrate how FiLiP can support the development of smart applications.

We have now also fixed the reference and caption issues.
