### 2024-07-26 d615adf
fix: caption and reference issues of the table
