### 2024-07-30 76ea36d
Making expressionLanguage attribute optional in ServiceGroup and Device. If it is set to None, it will be set to 'jexl'
