### 2024-07-30 8412915
Adding tests for setting expression language to None in ServiceGroup and Device
