+++ filip/models/ngsi_v2/iot.py
@@ -254,7 +254,7 @@ class ServiceGroup(BaseModel):
         default=ExpressionLanguage.JEXL,
         description="optional boolean value, to set expression language used "
                     "to compute expressions, possible values are: "
-                    "legacy or jexl, but legacy is deprecated."
+                    "legacy or jexl, but legacy is deprecated. If it is set None, jexl is used."
     )
     valid_expressionLanguage = field_validator("expressionLanguage")(validate_service_group_expression_language)
     explicitAttrs: Optional[bool] = Field(
@@ -325,7 +325,7 @@ class DeviceSettings(BaseModel):
         default=ExpressionLanguage.JEXL,
         description="optional boolean value, to set expression language used "
                     "to compute expressions, possible values are: "
-                    "legacy or jexl, but legacy is deprecated."
+                    "legacy or jexl, but legacy is deprecated. If it is set None, jexl is used."
     )
     valid_expressionLanguage = field_validator("expressionLanguage")(validate_device_expression_language)
     explicitAttrs: Optional[bool] = Field(
