+++ filip/models/ngsi_v2/iot.py
@@ -14,9 +14,10 @@ from filip.models.ngsi_v2.base import \
     BaseAttribute, \
     BaseValueAttribute, \
     BaseNameAttribute
-from filip.utils.validators import (validate_fiware_datatype_string_protect, validate_fiware_datatype_standard,
-                                    validate_jexl_expression, validate_device_expression_language,
-                                    validate_service_group_expression_language)
+from filip.utils.validators import (validate_fiware_datatype_string_protect,
+                                    validate_fiware_datatype_standard,
+                                    validate_jexl_expression,
+                                    validate_expression_language)
 
 logger = logging.getLogger()
 
@@ -256,7 +257,7 @@ class ServiceGroup(BaseModel):
                     "to compute expressions, possible values are: "
                     "legacy or jexl, but legacy is deprecated. If it is set None, jexl is used."
     )
-    valid_expressionLanguage = field_validator("expressionLanguage")(validate_service_group_expression_language)
+    valid_expressionLanguage = field_validator("expressionLanguage")(validate_expression_language)
     explicitAttrs: Optional[bool] = Field(
         default=False,
         description="optional boolean value, to support selective ignore "
@@ -327,7 +328,7 @@ class DeviceSettings(BaseModel):
                     "to compute expressions, possible values are: "
                     "legacy or jexl, but legacy is deprecated. If it is set None, jexl is used."
     )
-    valid_expressionLanguage = field_validator("expressionLanguage")(validate_device_expression_language)
+    valid_expressionLanguage = field_validator("expressionLanguage")(validate_expression_language)
     explicitAttrs: Optional[bool] = Field(
         default=False,
         description="optional boolean value, to support selective ignore "
