+++ setup.py
@@ -11,8 +11,8 @@ INSTALL_REQUIRES = ['aenum~=3.1.15',
                     'datamodel_code_generator[http]~=0.25.0',
                     'paho-mqtt~=2.0.0',
                     'pandas_datapackage_reader~=0.18.0',
-                    'pydantic~=2.5.2',
-                    'pydantic-settings~=2.0.0',
+                    'pydantic>=2.5.2,<2.7.0',
+                    'pydantic-settings>=2.0.0,<2.3.0',
                     'geojson_pydantic~=1.0.2',
                     'stringcase>=1.2.0',
                     'rdflib~=6.0.0',
@@ -25,7 +25,7 @@ INSTALL_REQUIRES = ['aenum~=3.1.15',
 
 SETUP_REQUIRES = INSTALL_REQUIRES.copy()
 
-VERSION = '0.4.1'
+VERSION = '0.5.0'
 
 setuptools.setup(
     name='filip',
@@ -40,7 +40,7 @@ setuptools.setup(
     download_url=f"https://github.com/RWTH-EBC/FiLiP/archive/refs/tags/v{VERSION}.tar.gz",
     project_urls={
         "Documentation":
-            "https://ebc.pages.rwth-aachen.de/EBC_all/github_ci/FiLiP/master/docs/index.html",
+            "https://rwth-ebc.github.io/FiLiP/master/docs/index.html",
         "Source":
             "https://github.com/RWTH-EBC/filip",
         "Download":
