+++ filip/__init__.py
@@ -4,4 +4,4 @@ filip-Module. See readme or documentation for more information.
 from filip.config import settings
 from filip.clients.ngsi_v2 import HttpClient
 
-__version__ = '0.4.1'
+__version__ = '0.5.0'
