+++ filip/models/ngsi_ld/context.py
@@ -523,13 +523,12 @@ class ContextLDEntity(ContextLDEntityKeyValues):
 
     model_config = ConfigDict(extra='allow', validate_default=True, validate_assignment=True)
 
-
     def model_dump(
         self,
         *args,
         by_alias: bool = True,
         **kwargs
-    ) -> dict[str, Any]:
+    ):
         return super().model_dump(*args, by_alias=by_alias, **kwargs)
 
     @field_validator("id")
