### 2024-07-31 b295598
chore: initial tenant and remove duplicated test
