### 2024-08-06 e06cf11
feat: Adding a custom exception class to make detailed exception handling inside different filip methods
