+++ filip/clients/exceptions.py
@@ -1,0 +1,17 @@
+"""
+Module for client specific exceptions
+"""
+import requests.models
+
+
+class BaseHttpClientException(Exception):
+    """
+    Base exception class for all HTTP clients. The response of a request will be available in the exception.
+
+    Args:
+        message (str): Error message
+        response (Response): Response object
+    """
+    def __init__(self, message: str, response: requests.models.Response):
+        super().__init__(message)
+        self.response = response
