### 2024-08-13 6fd2d03
added missing entity tests
Signed-off-by: iripiri <iris.m.koester@gmail.com>
