+++ tests/clients/test_ngsi_ld_entities.py
@@ -1,4 +1,3 @@
-import _json
 import unittest
 from pydantic import ValidationError
 
@@ -25,7 +24,6 @@ class TestEntities(unittest.TestCase):
         Cleanup entities from test server
         """
         entity_test_types = [ self.entity.type, self.entity_2.type ]
-        fiware_header = FiwareLDHeader()
         for entity_type in entity_test_types:
             entity_list = self.cb_client.get_entity_list(entity_type=entity_type)
             for entity in entity_list:
@@ -72,7 +70,18 @@ class TestEntities(unittest.TestCase):
             - limit(integer): Pagination limit
             - options(string): Options dictionary; Available values : keyValues, sysAttrs
         """
-        pass
+        entity_list = self.cb_client.get_entity_list()
+        self.assertEqual(len(entity_list), 0)
+
+        self.cb_client.post_entity(entity=self.entity)
+        entity_list_idpattern = self.cb_client.get_entity_list(id_pattern="urn:ngsi-ld:my*")
+        self.assertEqual(len(entity_list_idpattern), 1)
+        self.assertEqual(entity_list_idpattern[0].id, self.entity.id)
+
+        entity_list_attrs = self.cb_client.get_entity_list(attrs=["testtemperature"])
+        self.assertEqual(len(entity_list_attrs), 1)
+        self.assertEqual(entity_list_attrs[0].id, self.entity.id)
+
     
     def test_post_entity(self):
         """
@@ -134,7 +143,7 @@ class TestEntities(unittest.TestCase):
         self.assertEqual(len(entity_list), 1)
         self.assertEqual(entity_list[0].id, self.entity.id)
         self.assertEqual(entity_list[0].type, self.entity.type)
-        self.assertEqual(entity_list[0].testtemperature.value, self.entity.testtemperature.value)
+        self.assertEqual(entity_list[0].testtemperature, self.entity.testtemperature.value)
         
         """Test2"""
         self.entity_identical= self.entity.model_copy()
@@ -191,8 +200,19 @@ class TestEntities(unittest.TestCase):
         """Test1"""
         self.cb_client.post_entity(entity=self.entity)
         ret_entity = self.cb_client.get_entity(entity_id=self.entity.id)
+        ret_entity_with_type = self.cb_client.get_entity(entity_id=self.entity.id, entity_type=self.entity.type)
+        ret_entity_keyValues = self.cb_client.get_entity(entity_id=self.entity.id, options="keyValues")
+        ret_entity_sysAttrs = self.cb_client.get_entity(entity_id=self.entity.id, options="sysAttrs")
+
         self.assertEqual(ret_entity.id,self.entity.id)
         self.assertEqual(ret_entity.type,self.entity.type)
+        self.assertEqual(ret_entity_with_type.id,self.entity.id)
+        self.assertEqual(ret_entity_with_type.type,self.entity.type)
+        self.assertEqual(ret_entity_keyValues.id,self.entity.id)
+        self.assertEqual(ret_entity_keyValues.type,self.entity.type)
+        self.assertEqual(ret_entity_sysAttrs.id,self.entity.id)
+        self.assertEqual(ret_entity_sysAttrs.type,self.entity.type)
+        self.assertNotEqual(ret_entity_sysAttrs.createdAt, None)
 
         """Test2"""
         with self.assertRaises(requests.exceptions.HTTPError) as contextmanager:
@@ -319,7 +339,7 @@ class TestEntities(unittest.TestCase):
         self.cb_client.append_entity_attributes(self.entity)
         entity_list = self.cb_client.get_entity_list()
         for entity in entity_list:
-            self.assertEqual(first=entity.test_value.value, second=attr.value)
+            self.assertEqual(first=entity.test_value, second=attr.value)
         for entity in entity_list:
[…]
