### 2024-08-19 RWTH-EBC/FiLiP#317 @srmnitc
Changes to the paper
This is part of review at https://github.com/openjournals/joss-reviews/issues/6953

Some small change recommendations to the paper:

- [x] - authors: please add ORCids
- [x] - line 11: could be strengthened with a citation
- [x] - line 15-16: please add citations/links for NGSI-v2, Orion Content Broker, QuantumLeap
- [x] - table 1: could be nice to have the name of the library along with citation
- [x] - line 48: 'does not' -> 'do not'
- [x] - line 50: 'continue to' -> 'continues to'
- [x] - line 75: is _ needed in conversion factor?
- [x] - line 118: Is PID mentioned before? please check
- [x] - line 148: missing year in reference
- [x] - line 169: journal abbreviation, some our full names. Please choose one for consistency.

Thanks a lot for your efforts!
