### 2024-08-19 #6953 @srmnitc
@SBlechmann @djs0109 I have opened an issue with some small changes to the paper. As @djs0109 is out of office, we can wait until these are done, and then I can finish the rest of the steps. I will be out of office from 22.08-02.09, and will finish the rest of the steps once I am back. Meanwhile, I will also post here the remaining tasks that you need to do. Thanks once again for this submission and your efforts.
