### 2024-08-21 #6953 @djs0109
Hi @srmnitc , no problem! I will take care of these. I still have some questions regarding tasks 3-5. We have never made any release on those platforms before. Therefore, I think we could ignore these tasks. Could you please correct me or confirm my understanding?
