### 2024-08-21 #6953 @srmnitc
@djs0109 No, you have to make a release after review changes on github, but you also have to archive this release in platforms such as Zenodo or Figshare. These would then provide you a DOI which you have to post here. This will be included in the paper. The archiving step ensures that the published version would always be available.
