+++ joss/paper.bib
@@ -6,7 +6,7 @@
 	doi = {10.1088/1742-6596/1343/1/012053},
 	pages = {012053},
 	number = {1},
-	journaltitle = {J. Phys.: Conf. Ser.},
+	journaltitle = {Journal of Physics: Conference Series},
 	author = {Kümpel, Alexander and Storek, Thomas and Baranski, Marc and Schumacher, Markus and Müller, Dirk},
 	urldate = {2023-11-06},
 	date = {2019-11},
@@ -22,7 +22,7 @@
 	doi = {10.1088/1742-6596/1343/1/012063},
 	pages = {012063},
 	number = {1},
-	journaltitle = {J. Phys.: Conf. Ser.},
+	journaltitle = {Journal of Physics: Conference Series},
 	author = {Storek, T. and Lohmöller, J. and Kümpel, A. and Baranski, M. and Müller, D.},
 	urldate = {2023-11-06},
 	date = {2019-11},
@@ -116,6 +116,7 @@
 @misc{canterafonsecaFIWARENGSIV2Specification,
   title = {{{FIWARE-NGSI}} v2 {{Specification}}},
   author = {Cantera Fonseca, Jos{\'e} Manuel and M{\'a}rquez, Ferm{\'i}n Gal{\'a}n and Jacobs, Tobias},
+  year = {2018},
   urldate = {2023-12-15},
   abstract = {This specification defines the FIWARE-NGSI version 2 API. FIWARE-NGSI v2 is intended to manage the entire lifecycle of context information, including updates, queries, registrations, and subscriptions.}
 }
@@ -145,3 +146,16 @@
 	urldate = {2022-11-02},
 	date = {2023-01-01}
 }
+
+@article{araujo_performance_2019,
+	title = {Performance evaluation of {FIWARE}: {A} cloud-based {IoT} platform for smart cities},
+	volume = {132},
+	issn = {0743-7315},
+	url = {https://www.sciencedirect.com/science/article/pii/S0743731519300164},
+	doi = {10.1016/j.jpdc.2018.12.010},
+	journaltitle = {Journal of Parallel and Distributed Computing},
+	author = {Araujo, Victor and Mitra, Karan and Saguna, Saguna and Åhlund, Christer},
+	year = {2019},
+	pages = {250--261}
+}
+
