+++ joss/paper.md
@@ -9,15 +9,20 @@ tags:
   - Semantic
 authors:
   - name: Thomas Storek
+    orcid: 0000-0002-2652-1686
     affiliation: "1, 2"
   - name: Junsong Du
+    orcid: 0000-0003-2247-2423
     corresponding: true
     affiliation: 1 
   - name: Sebastian Blechmann
+    orcid: 0000-0002-8135-1843
     affiliation: 1
   - name: Rita Streblow
+    orcid: 0000-0001-7640-0930
     affiliation: 1
   - name: Dirk Müller
+    orcid: 0000-0002-6106-6607
     affiliation: 1
 affiliations:
   - name: Institute for Energy Efficient Buildings and Indoor Climate, RWTH Aachen University, Germany
@@ -30,10 +35,10 @@ bibliography: paper.bib
 
 # Summary
 
-FIWARE, an open-source initiative providing open-source software platform components and a set of standardized APIs (Application Programming Interfaces), has been instrumental in driving digitalization across various domains and research fields.
+FIWARE, an open-source initiative providing open-source software platform components and a set of standardized APIs (Application Programming Interfaces), has been instrumental in driving digitalization across various domains and research fields [@araujo_performance_2019].
 The use of FIWARE enables faster, easier, and cheaper developments of a wide range of IoT-enabled solutions [@fiware:2023]. 
 Over the years, the FIWARE Next-Generation-Service-Interface (NGSI-v2) API specification has been well-developed. 
-There is also a variety of reusable software components, so-called Generic Enablers (GEs), complying with NGSI-v2, including Orion Context Broker, IoT Agent, and QuantumLeap. 
+There is also a variety of reusable software components, so-called Generic Enablers ([GEs](https://github.com/FIWARE-GEs)), complying with NGSI-v2 [@canterafonsecaFIWARENGSIV2Specification], including [Orion Context Broker](https://github.com/telefonicaid/fiware-orion), [IoT Agent](https://github.com/telefonicaid/iotagent-node-lib), and [QuantumLeap](https://github.com/orchestracities/ngsi-timeseries-api?tab=readme-ov-file). 
 These components play a pivotal role in core context information management, IoT device connectivity, and time-series database interaction [@fiware:2023].
 
 In this context, we developed FiLiP, a Python Software Development Kit (SDK) designed to accelerate the development of services that interact with the aforementioned FIWARE GEs. 
@@ -51,22 +56,22 @@ Additionally, the specification has continuously evolved over time, necessitatin
 Existing FIWARE APIs client libraries can already provide some basic support but often fall short in several key areas.
 A comparison as shown in Table \ref{Comparison} reveals the shortage of various existing libraries, including limited support for multiple FIWARE APIs, lack of advanced validation and data models integration, and unsatisfactory usability and reliability.
 
-| Library                    | API Support                                               | Validation         | Datamodel            | Learning Support                         | Test Cases           |
-|----------------------------|-----------------------------------------------------------|--------------------|----------------------|------------------------------------------|----------------------|
-| [@QuantumLeapClient]       | TS                                                        | -                  | -                    | Readme file                              | 18                   |
-| [@fiware-ngsi-api]         | D, E                                                      | Basic    | Proprietary     | -                                        | -                    |
-| [@fiot-client-ngsi-python] | D<sup>*</sup>, E, S                                       | -                  | -                    | Readme file  | 31                   |
-| [@orion-python-client]     | E, S<sup>*</sup>                                          | -                  | -                    | Readme file                              | 1                    |
-| FiLiP                      | Device (D), Entity (E), Subscription (S), TimeSeries (TS) | Enhanced | Customizable         | Documentation, 12 examples, and 8 tutorials| 82 (84% coverage)    |
[…]
