### 2024-08-26 RWTH-EBC/FiLiP#317 @djs0109
Merged into [JOSS_submission](https://github.com/RWTH-EBC/FiLiP/tree/JOSS_submission)
