### 2024-08-26 55771b4
chore: add DOI for pydantic
