+++ joss/paper.bib
@@ -59,7 +59,8 @@
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/pydantic/pydantic}
+  url = {https://github.com/pydantic/pydantic},
+  doi = {10.5281/zenodo.10185654}
 }
 
 @misc{unitest,
