### 2024-08-26 #6953 @djs0109
@srmnitc I carefully checked the references again. The rest missing DOIs are either libraries or online documents that have no DOIs
