### 2024-08-27 2a8c55d
chore: integrate keyvalue update into existing functions
