### 2024-08-27 6c082ea
test: update attr with keyvalues
