### 2024-08-27 d45a17b
chore: example to create secure fiware header
