### 2024-09-03 0145681
tests: adapt property validation and unit tests
Subproperties are now checked recursively to deal with nested properties. But list properties apparently cause errors
