### 2024-09-04 94cfe61
tests: remove property list example
