### 2024-09-04 d0fa366
refactor: clarify property validator
