+++ filip/models/ngsi_ld/context.py
@@ -90,18 +90,12 @@ class ContextProperty(BaseModel):
         Returns:
             value
         """
-        if not value == "Property":
-            if value == "Relationship":
-                value == "Relationship"
-            elif value == "TemporalProperty":
-                value == "TemporalProperty"
-            else:
-                logging.warning(msg='NGSI_LD Properties must have type "Property"')
-                value = "Property"
+        valid_property_types = ["Property", "Relationship", "TemporalProperty"]
+        if value not in valid_property_types:
+            logging.warning(msg='NGSI_LD Properties must have type "Property"')
+            logging.warning(msg=f'Changing value from "{value}" to "Property"')
+            value = "Property"
         return value
-
-
-
 
 class NamedContextProperty(ContextProperty):
     """
@@ -500,7 +494,9 @@ class ContextLDEntity(ContextLDEntityKeyValues):
                  type: str,
                  **data):
         # There is currently no validation for extra fields
+        print('Data as is:\n',data)
         data.update(self._validate_attributes(data))
+        print('Data after updating:\n',data)
         super().__init__(id=id, type=type, **data)
 
     # TODO we should distinguish between context relationship
@@ -509,6 +505,7 @@ class ContextLDEntity(ContextLDEntityKeyValues):
     def _validate_attributes(cls, data: Dict):
         fields = set([field.validation_alias for (_, field) in cls.model_fields.items()] +
                      [field_name for field_name in cls.model_fields])
+        print('Fields: ',fields)
         fields.remove(None)
         # Initialize the attribute dictionary
         attrs = {}
@@ -516,12 +513,12 @@ class ContextLDEntity(ContextLDEntityKeyValues):
         # Iterate through the data
         for key, attr in data.items():
             # Check if the keyword is not already present in the fields
-            if key not in fields:
+            if key not in fields: # TODO why ignoring all in fields?
                 try:
                     attrs[key] = ContextGeoProperty.model_validate(attr)
                 except ValueError:
                     attrs[key] = ContextProperty.model_validate(attr)
-        return attrs
+        return attrs 
 
     model_config = ConfigDict(extra='allow', validate_default=True, validate_assignment=True)
 
@@ -552,40 +549,41 @@ class ContextLDEntity(ContextLDEntityKeyValues):
     #                 attrs[key] = ContextProperty.model_validate(attr)
     #     return attrs
     
-    def _validate_single_property(self, key, data, validity):
-        if key == 'type':
-            if data == DataTypeLD.RELATIONSHIP:
-                ContextRelationship.model_validate(data)
-            else:
-                ContextProperty.model_validate(data)
-        elif data is None or isinstance(data, (str, int, float)):
-            print('Skipping checking ',data,' because single value')
-            return validity
-        # elif isinstance(data, list):
-        #     for item in data:
-        #         validity = validity and self._validate_single_property(item, validity)
-        elif isinstance(data, dict):   
-            for attr_key, attr in data.items():                
-                validity = validity and self._validate_single_property(attr_key, attr, validity)
-        else:
-            raise NotImplementedError(
-            f"The property type ({type(data)}) for {data} is not implemented yet")
-        return validity
+    # def _validate_single_property(self, key, data, validity):
+    #     if key == 'type':
+    #         if data == DataTypeLD.RELATIONSHIP:
+    #             ContextRelationship.model_validate(data)
+    #         else:
+    #             ContextProperty.model_validate(data)
+    #     elif data is None or isinstance(data, (str, int, float)):
+    #         print('Skipping checking ',data,' because single value')
+    #         return validity
+    #     # elif isinstance(data, list):
+    #     #     for item in data:
[…]
