### 2024-09-04 801fefb
fix: rework batch operation
