### 2024-09-04 a20a457
chore: rename batch operation method
