+++ tests/clients/test_ngsi_ld_cb.py
@@ -101,18 +101,18 @@ class TestContextBroker(unittest.TestCase):
             entities_a = [ContextLDEntity(id=f"urn:ngsi-ld:test:{str(i)}",
                                         type=f'filip:object:TypeA') for i in
                           range(0, 1000)]
-            client.update(action_type=ActionTypeLD.CREATE, entities=entities_a)
+            client.entity_batch_operation(action_type=ActionTypeLD.CREATE, entities=entities_a)
             entities_b = [ContextLDEntity(id=f"urn:ngsi-ld:test:{str(i)}",
                                         type=f'filip:object:TypeB') for i in
                           range(1000, 2001)]
-            client.update(action_type=ActionTypeLD.CREATE, entities=entities_b)
+            client.entity_batch_operation(action_type=ActionTypeLD.CREATE, entities=entities_b)
             self.assertLessEqual(len(client.get_entity_list(limit=1)), 1)
             self.assertLessEqual(len(client.get_entity_list(limit=999)), 999)
             self.assertLessEqual(len(client.get_entity_list(limit=1001)), 1001)
             self.assertLessEqual(len(client.get_entity_list(limit=2001)), 2001)
 
-            client.update(action_type=ActionTypeLD.DELETE, entities=entities_a)
-            client.update(action_type=ActionTypeLD.DELETE, entities=entities_b)
+            client.entity_batch_operation(action_type=ActionTypeLD.DELETE, entities=entities_a)
+            client.entity_batch_operation(action_type=ActionTypeLD.DELETE, entities=entities_b)
 
     def aatest_entity_filtering(self):
         """
@@ -130,12 +130,12 @@ class TestContextBroker(unittest.TestCase):
                                         type=f'filip:object:TypeA') for i in
                           range(0, 5)]
 
-            client.update(action_type=ActionTypeLD.CREATE, entities=entities_a)
+            client.entity_batch_operation(action_type=ActionTypeLD.CREATE, entities=entities_a)
             entities_b = [ContextLDEntity(id=f"urn:ngsi-ld:TypeB:{str(i)}",
                                         type=f'filip:object:TypeB') for i in
                           range(6, 10)]
 
-            client.update(action_type=ActionTypeLD.CREATE, entities=entities_b)
+            client.entity_batch_operation(action_type=ActionTypeLD.CREATE, entities=entities_b)
 
             entities_all = client.get_entity_list()
             entities_by_id_pattern = client.get_entity_list(
@@ -162,9 +162,9 @@ class TestContextBroker(unittest.TestCase):
             with self.assertRaises(ValueError):
                 client.get_entity_list(response_format='not in AttrFormat')
 
-            client.update(action_type=ActionTypeLD.DELETE, entities=entities_a)
+            client.entity_batch_operation(action_type=ActionTypeLD.DELETE, entities=entities_a)
 
-            client.update(action_type=ActionTypeLD.DELETE, entities=entities_b)
+            client.entity_batch_operation(action_type=ActionTypeLD.DELETE, entities=entities_b)
 
     def test_entity_operations(self):
         """
@@ -274,31 +274,31 @@ class TestContextBroker(unittest.TestCase):
         entities = [ContextLDEntity(id=f"test:{i}",
                                     type=f'filip:object:TypeA') for i in
                     range(0, 1000)]
-        self.client.update(entities=entities, action_type=ActionTypeLD.CREATE)
+        self.client.entity_batch_operation(entities=entities, action_type=ActionTypeLD.CREATE)
         with self.assertRaises(RuntimeError):
             # the entity id must be unique
             entities = [ContextLDEntity(id=f"test:{i}",
                                         type=f'filip:object:TypeB') for i in
                         range(0, 1000)]
-            self.client.update(entities=entities, action_type=ActionTypeLD.CREATE)
[…]
