### 2024-09-10 #6953 @kyleniemeyer
@djs0109 we just need a few small edits before publishing:
- please add a space after "UN/CEFACT standards" before the citation
- At the beginning of the Use Case section, please combine the three citations into a single one using semicolons, like `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
